refactor(storage): Split redb.rs and test_suite.rs into directory modules
Split two large files into focused directory modules for better
maintainability. Pure code movement with zero behavioral changes.

- redb.rs (3,048 lines) → backends/redb/ (7 files)
- test_suite.rs (2,379 lines) → backends/test_suite/ (7 files)
- redb_config.rs moved to redb/config.rs
