--- crates/storage/src/backends/redb.rs
@@ -1,3048 +1,0 @@
-/// Redb backend implementation with snapshot isolation and ACID transactions.
-///
-/// This backend provides a pure Rust persistent key-value store using redb.
-/// It uses read snapshots for isolation and buffered writes for read-your-writes
-/// consistency within transactions.
-///
-/// # Features
-///
-/// - Pure Rust implementation (no C/C++ dependencies)
-/// - ACID transactions with snapshot isolation
-/// - Persistent storage with crash recovery
-/// - Single-writer model (matches Go DefraDB's LevelDB semantics)
-///
-/// # Platform Support
-///
-/// **NOTE**: This backend is NOT WASM-compatible due to redb's use of memory-mapped
-/// files and native filesystem operations. For WASM environments, use `MemoryStore`
-/// instead, or implement a browser-specific backend using IndexedDB/OPFS.
-///
-/// # MVCC Behavior
-///
-/// When a transaction is created, it opens a redb `ReadTransaction` which provides
-/// zero-copy MVCC snapshot isolation. All reads within the transaction see the
-/// database state as it was at creation time, regardless of concurrent writes.
-///
-/// # Async Callback Lifecycle
-///
-/// Transaction callbacks follow fire-and-forget semantics (matching Go DefraDB):
-///
-/// - **Sync callbacks**: Executed inline during commit/discard, blocking until complete
-/// - **Async callbacks on commit**: Awaited during commit, blocking return until complete
-/// - **Async callbacks on discard**: Spawned as background tasks (fire-and-forget)
-///
-/// **Important**: Async discard callbacks may not complete if the process exits
-/// before they finish. Callers requiring completion guarantees should use
-/// `tokio::task::JoinSet`, `tokio_util::task::TaskTracker`, or similar
-/// synchronization, or prefer `commit()` over `discard()` when async cleanup
-/// is critical.
-///
-/// # Use Cases
-///
-/// - Production deployments on native platforms (Linux, macOS, Windows)
-/// - Embedded applications with filesystem access
-///
-/// # Example
-///
-/// ```ignore
-/// use storage::backends::redb::RedbStore;
-/// use storage::corekv::{Store, Reader, Writer};
-///
-/// let store = RedbStore::open("/path/to/db")?;
-/// let mut txn = store.new_txn(false).await?;
-/// txn.set(b"key", b"value").await?;
-/// txn.commit().await?;
-/// ```
-use async_trait::async_trait;
-use parking_lot::Mutex;
-use redb::{Database, ReadTransaction, ReadableTable, TableDefinition};
-use std::collections::BTreeMap;
-use std::collections::HashSet;
-use std::ops::Bound;
-use std::path::Path;
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::sync::Arc;
-use tokio::sync::RwLock;
-
-use super::shared::ConflictTracker;
-
-use crate::corekv::{
-    AsyncTxnCallback, Dropable, Error, IterOptions, Iterator, KvPair, Reader, Result, Store, Txn,
-    TxnCallback, Writer,
-};
-
-use super::redb_config::{DurabilityMode, RedbStoreOptions};
-
-/// Table definition for the main key-value store.
-const KV_TABLE: TableDefinition<&[u8], &[u8]> = TableDefinition::new("kv");
-
-/// Redb-backed key-value store.
-///
-/// This store wraps a redb Database instance and provides MVCC transaction
-/// support through snapshots and buffered writes.
-///
-/// # Active Transaction Tracking
-///
-/// The store tracks the number of active transactions. When closing, the store
-/// will reject new transactions and wait for existing ones to complete.
-pub struct RedbStore {
-    db: Arc<Database>,
-    closed: Arc<RwLock<bool>>,
-    /// Count of active transactions (for graceful shutdown)
-    active_txn_count: Arc<AtomicUsize>,
-    /// Close timeout duration
-    close_timeout: std::time::Duration,
-    /// Database file path (for error messages)
-    db_path: std::path::PathBuf,
-    /// Conflict tracker for write-write conflict detection
-    conflict_tracker: Arc<ConflictTracker>,
-    /// Durability mode for write transactions
-    durability: DurabilityMode,
-}
-
-impl RedbStore {
-    /// Open a redb database at the specified path with default options.
-    ///
-    /// If the database doesn't exist, it will be created.
-    /// If the path is a directory, creates `data.redb` inside it.
-    ///
-    /// # Arguments
-    ///
-    /// * `path` - Path to the database file or directory
-    ///
-    /// # Returns
-    ///
-    /// * `Ok(RedbStore)` on success
-    /// * `Err(Error)` if the database cannot be opened
-    pub fn open<P: AsRef<Path>>(path: P) -> Result<Self> {
-        Self::open_with_options(path, RedbStoreOptions::default())
-    }
-
-    /// Open a redb database at the specified path with custom options.
-    ///
-    /// If the database doesn't exist, it will be created.
-    /// If the path is a directory, creates `data.redb` inside it.
-    ///
-    /// # Arguments
-    ///
-    /// * `path` - Path to the database file or directory
-    /// * `opts` - Configuration options for the database
-    ///
-    /// # Returns
-    ///
-    /// * `Ok(RedbStore)` on success
-    /// * `Err(Error)` if the database cannot be opened
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// use storage::backends::{RedbStore, RedbStoreOptions};
-    ///
-    /// let opts = RedbStoreOptions::new()
-    ///     .with_cache_size(64 * 1024 * 1024); // 64MB cache
-    ///
-    /// let store = RedbStore::open_with_options("/path/to/db", opts)?;
-    /// ```
-    pub fn open_with_options<P: AsRef<Path>>(path: P, opts: RedbStoreOptions) -> Result<Self> {
-        let path = path.as_ref();
-        let db_path = if path.is_dir() {
-            path.join("data.redb")
-        } else {
-            path.to_path_buf()
-        };
-
-        let mut builder = redb::Builder::new();
-        if let Some(cache_size) = opts.cache_size() {
-            builder.set_cache_size(cache_size);
-        }
-
-        // Open database with path context in error messages
-        let db = builder.create(&db_path).map_err(|e| {
-            // Add file path context to common errors
-            match &e {
-                redb::DatabaseError::DatabaseAlreadyOpen => {
-                    tracing::warn!(db_path = %db_path.display(), "Database is locked by another process");
-                    Error::Backend(format!(
-                        "database '{}' is locked by another process. \
-                         Troubleshooting: (1) check for other running processes using 'lsof {}', \
-                         (2) if no process found, the lock file may be stale - check for .lock files, \
-                         (3) ensure previous instance shut down cleanly",
-                        db_path.display(),
-                        db_path.display()
-                    ))
-                }
-                redb::DatabaseError::UpgradeRequired(version) => {
-                    tracing::warn!(
-                        db_path = %db_path.display(),
-                        version = version,
-                        "Database file format upgrade required"
-                    );
-                    Error::Backend(format!(
-                        "database '{}' uses file format version {} which requires upgrade. \
-                         To upgrade: (1) backup your database first, \
-                         (2) use 'redb-cli upgrade {}' or equivalent tool, \
-                         (3) see redb documentation for migration details",
-                        db_path.display(),
-                        version,
-                        db_path.display()
-                    ))
-                }
-                _ => {
-                    tracing::error!(
-                        db_path = %db_path.display(),
-                        error = %e,
-                        "Failed to open database"
-                    );
-                    e.into()
-                }
-            }
-        })?;
-
-        // Ensure the KV table exists by opening a write transaction
-        {
-            let write_txn = db.begin_write()?;
-            let _ = write_txn.open_table(KV_TABLE)?;
-            write_txn.commit()?;
-        }
-
-        Ok(Self {
-            db: Arc::new(db),
-            closed: Arc::new(RwLock::new(false)),
-            active_txn_count: Arc::new(AtomicUsize::new(0)),
-            close_timeout: opts.close_timeout(),
-            db_path,
-            conflict_tracker: Arc::new(ConflictTracker::new()),
-            durability: opts.durability(),
-        })
-    }
-
-    /// Get the current count of active transactions.
-    pub fn active_transaction_count(&self) -> usize {
-        self.active_txn_count.load(Ordering::SeqCst)
-    }
-
-    /// Get the database file path.
-    pub fn db_path(&self) -> &std::path::Path {
-        &self.db_path
-    }
-
-    /// Check database integrity.
-    ///
-    /// This performs a read-only scan of the database to verify that all data
-    /// is readable and the internal structure is consistent. It does NOT repair
-    /// any issues found.
-    ///
-    /// # Returns
-    ///
-    /// * `Ok(IntegrityReport)` with detailed results of the check
-    /// * `Err(Error)` if the check could not be completed
-    ///
-    /// # Note
-    ///
-    /// This operation may be slow on large databases as it reads all data.
-    pub fn check_integrity(&self) -> Result<IntegrityReport> {
-        // Attempt to read all keys to verify data integrity
-        let read_txn = self.db.begin_read()?;
-
-        let table = match read_txn.open_table(KV_TABLE) {
-            Ok(t) => t,
-            Err(redb::TableError::TableDoesNotExist(_)) => {
-                // Empty database is valid
-                return Ok(IntegrityReport {
-                    is_valid: true,
-                    total_keys: 0,
-                    error_count: 0,
-                    first_error: None,
-                });
-            }
-            Err(e) => {
-                tracing::error!(error = %e, "Integrity check failed: cannot open table");
-                return Err(e.into());
-            }
-        };
-
-        let mut key_count = 0u64;
-        let mut error_count = 0u64;
-        let mut first_error: Option<String> = None;
-
-        let range = match table.range::<&[u8]>(..) {
-            Ok(r) => r,
-            Err(e) => {
-                tracing::error!(error = %e, "Integrity check failed: cannot create range iterator");
-                return Err(e.into());
-            }
-        };
-
-        for result in range {
-            match result {
-                Ok((key, value)) => {
-                    // Try to read the key and value to verify they're accessible
-                    let _ = key.value();
-                    let _ = value.value();
-                    key_count += 1;
-                }
-                Err(e) => {
-                    let error_msg = e.to_string();
-                    tracing::error!(
-                        error = %e,
-                        keys_checked = key_count,
-                        "Integrity check found error while reading key"
-                    );
-                    if first_error.is_none() {
-                        first_error = Some(error_msg);
-                    }
-                    error_count += 1;
-                }
-            }
-        }
-
-        let is_valid = error_count == 0;
-        if !is_valid {
-            tracing::warn!(
-                total_keys = key_count,
-                errors = error_count,
-                "Integrity check completed with errors"
-            );
-        } else {
-            tracing::info!(total_keys = key_count, "Integrity check passed");
-        }
-
-        Ok(IntegrityReport {
-            is_valid,
-            total_keys: key_count,
-            error_count,
-            first_error,
-        })
-    }
-
-    /// Check if the store is closed.
-    async fn is_closed(&self) -> bool {
-        *self.closed.read().await
-    }
-}
-
-#[async_trait]
-impl Store for RedbStore {
-    async fn new_txn(&self, readonly: bool) -> Result<Box<dyn Txn>> {
-        // Check closed status and increment count atomically to prevent TOCTOU race.
-        // By holding the read lock while incrementing, we ensure close() will see
-        // our incremented count if it runs concurrently.
-        {
-            let closed = self.closed.read().await;
-            if *closed {
-                return Err(Error::DBClosed);
-            }
-            self.active_txn_count.fetch_add(1, Ordering::SeqCst);
-        }
-
-        // Use a local guard to ensure count is decremented on panic or early return.
-        // The guard is defused (set to true) only when the transaction is fully constructed.
-        struct NewTxnGuard<'a>(&'a AtomicUsize, bool);
-        impl Drop for NewTxnGuard<'_> {
-            fn drop(&mut self) {
-                if !self.1 {
-                    self.0.fetch_sub(1, Ordering::SeqCst);
-                }
-            }
-        }
-        let mut guard = NewTxnGuard(&self.active_txn_count, false);
-
-        // Record version before taking snapshot for conflict detection
-        let read_version = self.conflict_tracker.current_version();
-
-        // Use redb's MVCC ReadTransaction for snapshot isolation (O(1) creation)
-        let read_txn = self.db.begin_read()?;
-
-        // Defuse the guard - transaction will manage its own count via its Drop impl
-        guard.1 = true;
-
-        Ok(Box::new(RedbTxn {
-            db: Arc::clone(&self.db),
-            active_txn_count: Arc::clone(&self.active_txn_count),
-            conflict_tracker: Arc::clone(&self.conflict_tracker),
-            read_version,
-            read_txn,
-            pending: Mutex::new(BTreeMap::new()),
-            readonly,
-            durability: self.durability,
-            discarded: Mutex::new(false),
-            committed: Mutex::new(false),
-            on_success: Mutex::new(Vec::new()),
-            on_success_async: Mutex::new(Vec::new()),
-            on_error: Mutex::new(Vec::new()),
-            on_error_async: Mutex::new(Vec::new()),
-            on_discard: Mutex::new(Vec::new()),
-            on_discard_async: Mutex::new(Vec::new()),
-        }))
-    }
-
-    async fn close(&self) -> Result<()> {
-        // Mark as closed first to prevent new transactions.
-        // If already closed, return early (idempotent behavior).
-        {
-            let mut closed = self.closed.write().await;
-            if *closed {
-                return Ok(());
-            }
-            *closed = true;
-        }
-
-        // Wait for active transactions to complete (with timeout)
-        let active = self.active_txn_count.load(Ordering::SeqCst);
-        if active > 0 {
-            tracing::info!(
-                active_transactions = active,
-                db_path = %self.db_path.display(),
-                "Store closing with active transactions - waiting for completion"
-            );
-
-            // Poll for transactions to complete (uses configurable timeout)
-            let start = std::time::Instant::now();
-            let timeout = self.close_timeout;
-            while self.active_txn_count.load(Ordering::SeqCst) > 0 {
-                if start.elapsed() > timeout {
-                    let remaining = self.active_txn_count.load(Ordering::SeqCst);
-                    tracing::error!(
-                        remaining_transactions = remaining,
-                        timeout_secs = timeout.as_secs(),
-                        db_path = %self.db_path.display(),
-                        "Failed to close store - transactions still active after timeout"
-                    );
-                    return Err(Error::Other(format!(
-                        "Close timeout: {} transaction(s) still active after {}s (db: {}). \
-                         Possible causes: (1) Transactions not calling commit()/discard() - check for missing cleanup, \
-                         (2) Long-running I/O operations - transactions may still be processing large commits or snapshots. \
-                         Use RedbStoreOptions::with_close_timeout() to increase timeout if needed.",
-                        remaining,
-                        timeout.as_secs(),
-                        self.db_path.display()
-                    )));
-                }
-                tokio::time::sleep(std::time::Duration::from_millis(10)).await;
-            }
-        }
-
-        Ok(())
-    }
-}
-
-#[async_trait]
-impl Dropable for RedbStore {
-    async fn drop_all(&self) -> Result<()> {
-        if self.is_closed().await {
-            return Err(Error::DBClosed);
-        }
-
-        // Open a write transaction and delete all entries
-        let write_txn = self.db.begin_write()?;
-        {
-            let mut table = write_txn.open_table(KV_TABLE)?;
-            // Collect all keys first to avoid borrow issues
-            let keys: Vec<Vec<u8>> = {
-                let range = table.range::<&[u8]>(..)?;
-                range
-                    .map(|result| result.map(|(k, _)| k.value().to_vec()))
-                    .collect::<std::result::Result<Vec<_>, _>>()?
-            };
-            // Delete all keys
-            for key in keys {
-                table.remove(key.as_slice())?;
-            }
-        }
-        write_txn.commit()?;
-
-        Ok(())
-    }
-}
-
-/// Compute the start and end bounds for a BTreeMap range query from IterOptions.
-///
-/// This optimizes iteration by using the underlying data structure's range
-/// capabilities instead of filtering after iteration.
-fn compute_range_bounds(opts: &IterOptions) -> (Bound<Vec<u8>>, Bound<Vec<u8>>) {
-    // Determine start bound
-    let start_bound = match (opts.prefix(), opts.start()) {
-        (Some(prefix), Some(start)) => {
-            // Use whichever is greater
-            if prefix > start {
-                Bound::Included(prefix.to_vec())
-            } else {
-                Bound::Included(start.to_vec())
-            }
-        }
-        (Some(prefix), None) => Bound::Included(prefix.to_vec()),
-        (None, Some(start)) => Bound::Included(start.to_vec()),
-        (None, None) => Bound::Unbounded,
-    };
-
-    // Determine end bound
-    let end_bound = match (opts.prefix(), opts.end()) {
-        (Some(prefix), Some(end)) => {
-            // Compute prefix end (prefix with last byte incremented)
-            let prefix_end = prefix_to_end_bound(prefix);
-            // Use whichever is smaller
-            if let Some(pe) = prefix_end {
-                if pe.as_slice() < end {
-                    Bound::Excluded(pe)
-                } else {
-                    Bound::Excluded(end.to_vec())
-                }
-            } else {
-                Bound::Excluded(end.to_vec())
-            }
-        }
-        (Some(prefix), None) => {
-            match prefix_to_end_bound(prefix) {
-                Some(end) => Bound::Excluded(end),
-                None => Bound::Unbounded, // Prefix is all 0xFF bytes
-            }
-        }
-        (None, Some(end)) => Bound::Excluded(end.to_vec()),
-        (None, None) => Bound::Unbounded,
-    };
-
-    (start_bound, end_bound)
-}
-
-/// Convert a `Bound<Vec<u8>>` to `Bound<&[u8]>` for redb range queries.
-fn bound_as_ref(bound: &Bound<Vec<u8>>) -> Bound<&[u8]> {
-    match bound {
-        Bound::Included(v) => Bound::Included(v.as_slice()),
-        Bound::Excluded(v) => Bound::Excluded(v.as_slice()),
-        Bound::Unbounded => Bound::Unbounded,
-    }
-}
-
-/// Compute the exclusive end bound for a prefix.
-///
-/// Given a prefix like "foo", returns "fop" (the first key that doesn't match the prefix).
-/// Returns None if the prefix is empty or all 0xFF bytes (meaning iteration should go to the end).
-fn prefix_to_end_bound(prefix: &[u8]) -> Option<Vec<u8>> {
-    // Empty prefix matches all keys - no end bound needed
-    if prefix.is_empty() {
-        return None;
-    }
-
-    let mut end = prefix.to_vec();
-    // Increment the last byte, handling overflow
-    while let Some(last) = end.pop() {
-        if last < 0xFF {
-            end.push(last + 1);
-            return Some(end);
-        }
-        // If the byte was 0xFF, we popped it and try the next one
-    }
-    // All bytes were 0xFF
-    None
-}
-
-/// Redb transaction with snapshot isolation and buffered writes.
-///
-/// Transactions maintain a snapshot of the store at creation time and track
-/// pending changes. Changes are applied atomically on commit.
-///
-/// # Drop Safety
-///
-/// If a transaction is dropped without calling `commit()` or `discard()`,
-/// the Drop implementation will:
-/// - Decrement the active transaction count (preventing store close hangs)
-/// - Log a warning about the improper cleanup
-///
-/// This is a safety net - callers should always explicitly commit or discard.
-struct RedbTxn {
-    /// Reference to the redb database
-    db: Arc<Database>,
-
-    /// Reference to the active transaction counter (for decrement on complete)
-    active_txn_count: Arc<AtomicUsize>,
-
-    /// Conflict tracker for write-write conflict detection
-    conflict_tracker: Arc<ConflictTracker>,
-
-    /// Version at which this transaction's snapshot was taken
-    read_version: u64,
-
-    /// Redb MVCC read transaction for snapshot isolation (O(1) creation)
-    read_txn: ReadTransaction,
-
-    /// Pending changes (Some(value) = set, None = delete)
-    pending: Mutex<BTreeMap<Vec<u8>, Option<Vec<u8>>>>,
-
-    /// Whether this is a read-only transaction
-    readonly: bool,
-
-    /// Durability mode for write commits
-    durability: DurabilityMode,
-
-    /// Whether the transaction has been discarded
-    discarded: Mutex<bool>,
-
-    /// Whether the transaction has been committed
-    committed: Mutex<bool>,
-
-    /// Callbacks for successful commit
-    on_success: Mutex<Vec<TxnCallback>>,
-    on_success_async: Mutex<Vec<AsyncTxnCallback>>,
-
-    /// Callbacks for failed commit
-    on_error: Mutex<Vec<TxnCallback>>,
-    on_error_async: Mutex<Vec<AsyncTxnCallback>>,
-
-    /// Callbacks for discard
-    on_discard: Mutex<Vec<TxnCallback>>,
-    on_discard_async: Mutex<Vec<AsyncTxnCallback>>,
-}
-
-impl Drop for RedbTxn {
-    fn drop(&mut self) {
-        // Always decrement the active transaction count
-        self.active_txn_count.fetch_sub(1, Ordering::SeqCst);
-
-        // Log warning if dropped without explicit commit/discard
-        let was_committed = *self.committed.lock();
-        let was_discarded = *self.discarded.lock();
-        if !was_committed && !was_discarded {
-            // Count skipped callbacks to include in warning
-            let skipped_discard = self.on_discard.lock().len();
-            let skipped_discard_async = self.on_discard_async.lock().len();
-            let total_skipped = skipped_discard + skipped_discard_async;
-
-            if total_skipped > 0 {
-                tracing::warn!(
-                    skipped_callbacks = total_skipped,
-                    "Transaction dropped without commit() or discard() - \
-                     this may indicate a bug. Pending changes were lost and \
-                     {} registered discard callback(s) were NOT executed.",
-                    total_skipped
-                );
-            } else {
-                tracing::warn!(
-                    "Transaction dropped without commit() or discard() - \
-                     this may indicate a bug. Pending changes were lost."
-                );
-            }
-        }
-    }
-}
-
-/// Callback counts for monitoring transaction callback accumulation.
-#[derive(Debug, Clone, Default)]
-pub struct CallbackCounts {
-    /// Number of synchronous on_success callbacks registered
-    pub on_success: usize,
-    /// Number of asynchronous on_success callbacks registered
-    pub on_success_async: usize,
-    /// Number of synchronous on_error callbacks registered
-    pub on_error: usize,
-    /// Number of asynchronous on_error callbacks registered
-    pub on_error_async: usize,
-    /// Number of synchronous on_discard callbacks registered
-    pub on_discard: usize,
-    /// Number of asynchronous on_discard callbacks registered
-    pub on_discard_async: usize,
-}
-
-impl CallbackCounts {
-    /// Total number of callbacks registered across all types.
-    pub fn total(&self) -> usize {
-        self.on_success
-            + self.on_success_async
-            + self.on_error
-            + self.on_error_async
-            + self.on_discard
-            + self.on_discard_async
-    }
-}
-
-/// Report from database integrity check.
-///
-/// Provides detailed information about the integrity check results,
-/// including total keys scanned and any errors encountered.
-#[derive(Debug, Clone)]
-pub struct IntegrityReport {
-    /// Whether the integrity check passed (no errors found).
-    pub is_valid: bool,
-    /// Total number of keys scanned during the check.
-    pub total_keys: u64,
-    /// Number of errors encountered during the check.
-    pub error_count: u64,
-    /// First error message encountered, if any (for debugging).
-    pub first_error: Option<String>,
-}
-
-impl IntegrityReport {
-    /// Check if the database passed the integrity check.
-    pub fn is_valid(&self) -> bool {
-        self.is_valid
-    }
-}
-
-impl RedbTxn {
-    /// Get the current count of registered callbacks.
-    ///
-    /// This is useful for monitoring callback accumulation in long-lived
-    /// transactions and detecting potential memory pressure.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let txn = store.new_txn(false).await?;
-    /// // ... register some callbacks ...
-    /// let counts = txn.callback_counts();
-    /// if counts.total() > 100 {
-    ///     tracing::warn!("Transaction has {} callbacks registered", counts.total());
-    /// }
-    /// ```
-    #[allow(dead_code)] // Part of public API - used externally for monitoring
-    pub fn callback_counts(&self) -> CallbackCounts {
-        CallbackCounts {
-            on_success: self.on_success.lock().len(),
-            on_success_async: self.on_success_async.lock().len(),
-            on_error: self.on_error.lock().len(),
-            on_error_async: self.on_error_async.lock().len(),
-            on_discard: self.on_discard.lock().len(),
-            on_discard_async: self.on_discard_async.lock().len(),
-        }
-    }
-
-    /// Get a value, checking pending changes first, then the read transaction.
-    fn get_internal(&self, key: &[u8]) -> Result<Option<Vec<u8>>> {
-        // Check pending changes first
-        let pending = self.pending.lock();
-        if let Some(pending_value) = pending.get(key) {
-            return Ok(pending_value.clone());
-        }
-        drop(pending);
-
-        // Fall back to redb ReadTransaction
-        let table = match self.read_txn.open_table(KV_TABLE) {
-            Ok(t) => t,
-            Err(redb::TableError::TableDoesNotExist(_)) => return Ok(None),
-            Err(e) => return Err(e.into()),
-        };
-        match table.get(key)? {
-            Some(value) => Ok(Some(value.value().to_vec())),
-            None => Ok(None),
-        }
-    }
-
-    /// Check if a key exists.
-    fn has_internal(&self, key: &[u8]) -> Result<bool> {
-        // Check pending changes first
-        let pending = self.pending.lock();
-        if let Some(pending_value) = pending.get(key) {
-            return Ok(pending_value.is_some());
-        }
-        drop(pending);
-
-        // Fall back to redb ReadTransaction
-        let table = match self.read_txn.open_table(KV_TABLE) {
-            Ok(t) => t,
-            Err(redb::TableError::TableDoesNotExist(_)) => return Ok(false),
-            Err(e) => return Err(e.into()),
-        };
-        Ok(table.get(key)?.is_some())
-    }
-
-    /// Execute sync callbacks with panic protection.
-    ///
-    /// Returns the number of callbacks that panicked.
-    fn execute_callbacks(callbacks: Vec<TxnCallback>) -> usize {
-        let total = callbacks.len();
-        let mut failed = 0;
-
-        for (i, callback) in callbacks.into_iter().enumerate() {
-            let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(callback));
-            if let Err(panic_info) = result {
-                failed += 1;
-                tracing::error!(
-                    callback_index = i,
-                    panic = ?panic_info,
-                    "Transaction callback panicked - continuing with remaining callbacks"
-                );
-            }
-        }
-
-        if failed > 0 {
-            tracing::warn!(
-                total_callbacks = total,
-                failed_callbacks = failed,
-                "Sync callback execution completed with failures"
-            );
-        }
-
-        failed
-    }
-
-    /// Execute async callbacks with panic protection.
-    ///
-    /// Returns the number of callbacks that panicked.
-    async fn execute_async_callbacks(callbacks: Vec<AsyncTxnCallback>) -> usize {
-        use futures::FutureExt;
-
-        let total = callbacks.len();
-        let mut failed = 0;
-
-        for (i, callback) in callbacks.into_iter().enumerate() {
-            let future = callback();
-            let result = std::panic::AssertUnwindSafe(future).catch_unwind().await;
-            if let Err(panic_info) = result {
-                failed += 1;
-                tracing::error!(
-                    callback_index = i,
-                    panic = ?panic_info,
-                    "Async callback panicked during execution - continuing with remaining callbacks"
-                );
-            }
-        }
-
-        if failed > 0 {
-            tracing::warn!(
-                total_callbacks = total,
-                failed_callbacks = failed,
-                "Async callback execution completed with failures"
-            );
-        }
-
-        failed
-    }
-}
-
-#[async_trait]
-impl Reader for RedbTxn {
-    async fn get(&self, key: &[u8]) -> Result<Option<Vec<u8>>> {
-        if *self.discarded.lock() {
-            return Err(Error::DiscardedTxn);
-        }
-
-        if key.is_empty() {
-            return Err(Error::EmptyKey);
-        }
-
-        self.get_internal(key)
-    }
-
-    async fn has(&self, key: &[u8]) -> Result<bool> {
-        if *self.discarded.lock() {
-            return Err(Error::DiscardedTxn);
-        }
-
-        if key.is_empty() {
-            return Err(Error::EmptyKey);
-        }
-
-        self.has_internal(key)
-    }
-
-    async fn get_size(&self, key: &[u8]) -> Result<Option<usize>> {
-        if *self.discarded.lock() {
-            return Err(Error::DiscardedTxn);
-        }
-
-        if key.is_empty() {
-            return Err(Error::EmptyKey);
-        }
-
-        Ok(self.get_internal(key)?.map(|v| v.len()))
-    }
-
-    async fn iterator(&self, opts: IterOptions) -> Result<Box<dyn Iterator>> {
-        if *self.discarded.lock() {
-            return Err(Error::DiscardedTxn);
-        }
-
-        // Compute the effective range bounds for efficient range queries
-        let (start_bound, end_bound) = compute_range_bounds(&opts);
-
-        // Helper to check prefix
-        let matches_prefix =
-            |key: &[u8]| -> bool { opts.prefix().is_none_or(|p| key.starts_with(p)) };
-
-        // Read matching items from the redb ReadTransaction (only matching range)
-        let snapshot_items: Vec<(Vec<u8>, Vec<u8>)> = match self.read_txn.open_table(KV_TABLE) {
-            Ok(table) => {
-                let range =
-                    table.range::<&[u8]>((bound_as_ref(&start_bound), bound_as_ref(&end_bound)))?;
-                let mut items = Vec::new();
-                for result in range {
-                    let (key, value) = result?;
-                    let k = key.value().to_vec();
-                    if matches_prefix(&k) {
-                        items.push((k, value.value().to_vec()));
-                    }
-                }
-                items
-            }
-            Err(redb::TableError::TableDoesNotExist(_)) => Vec::new(),
-            Err(e) => return Err(e.into()),
-        };
-
-        // Extract pending items into Vec (sorted by BTreeMap, with Option for deletions)
-        let pending = self.pending.lock();
-        let pending_items: Vec<(Vec<u8>, Option<Vec<u8>>)> = pending
-            .range((start_bound, end_bound))
-            .filter(|(k, _)| matches_prefix(k))
-            .map(|(k, v)| (k.clone(), v.clone()))
-            .collect();
-
-        Ok(Box::new(MergingIterator::new(
-            snapshot_items,
-            pending_items,
-            opts,
-        )))
-    }
-}
-
-#[async_trait]
-impl Writer for RedbTxn {
-    async fn set(&mut self, key: &[u8], value: &[u8]) -> Result<()> {
-        if *self.discarded.lock() {
-            return Err(Error::DiscardedTxn);
-        }
-
-        if self.readonly {
-            return Err(Error::ReadOnlyTxn);
-        }
-
-        if key.is_empty() {
-            return Err(Error::EmptyKey);
-        }
-
-        self.pending
-            .lock()
-            .insert(key.to_vec(), Some(value.to_vec()));
-        Ok(())
-    }
-
-    async fn delete(&mut self, key: &[u8]) -> Result<()> {
-        if *self.discarded.lock() {
-            return Err(Error::DiscardedTxn);
-        }
-
-        if self.readonly {
-            return Err(Error::ReadOnlyTxn);
-        }
-
-        if key.is_empty() {
-            return Err(Error::EmptyKey);
-        }
-
-        self.pending.lock().insert(key.to_vec(), None);
-        Ok(())
-    }
-}
-
-#[async_trait]
-impl Txn for RedbTxn {
-    async fn commit(self: Box<Self>) -> Result<()> {
-        // Note: active_txn_count is decremented by Drop impl when self is dropped
-        // at the end of this function (on any exit path).
-
-        if *self.discarded.lock() {
-            tracing::warn!("Attempted to commit a discarded transaction");
-            let on_error = std::mem::take(&mut *self.on_error.lock());
-            let on_error_async = std::mem::take(&mut *self.on_error_async.lock());
-            Self::execute_callbacks(on_error);
-            Self::execute_async_callbacks(on_error_async).await;
-            return Err(Error::DiscardedTxn);
-        }
-
-        if *self.committed.lock() {
-            tracing::warn!("Attempted to commit an already committed transaction");
-            return Err(Error::Other("Transaction already committed".into()));
-        }
-
-        // Take pending changes (avoids clone — commit consumes self via Box<Self>)
-        let pending = std::mem::take(&mut *self.pending.lock());
-
-        // Check for write-write conflicts before applying
-        if !pending.is_empty() {
-            let write_set: HashSet<Vec<u8>> = pending.keys().cloned().collect();
-            if let Err(e) = self
-                .conflict_tracker
-                .check_and_record(self.read_version, write_set)
-            {
-                let on_error = std::mem::take(&mut *self.on_error.lock());
-                let on_error_async = std::mem::take(&mut *self.on_error_async.lock());
-                Self::execute_callbacks(on_error);
-                Self::execute_async_callbacks(on_error_async).await;
-                return Err(e);
-            }
-        }
-
-        // Apply pending changes to the database if there are any
-        if !pending.is_empty() {
-            let mut write_txn = match self.db.begin_write() {
-                Ok(txn) => txn,
-                Err(e) => {
-                    tracing::error!(
-                        error = %e,
-                        pending_changes = pending.len(),
-                        "Failed to begin write transaction during commit"
-                    );
-                    let on_error = std::mem::take(&mut *self.on_error.lock());
-                    let on_error_async = std::mem::take(&mut *self.on_error_async.lock());
-                    Self::execute_callbacks(on_error);
-                    Self::execute_async_callbacks(on_error_async).await;
-                    return Err(e.into());
-                }
-            };
-
-            write_txn.set_durability(match self.durability {
-                DurabilityMode::Immediate => redb::Durability::Immediate,
-                DurabilityMode::Eventual => redb::Durability::Eventual,
-            });
-
-            {
-                let mut table = match write_txn.open_table(KV_TABLE) {
-                    Ok(t) => t,
-                    Err(e) => {
-                        tracing::error!(
-                            error = %e,
-                            "Failed to open KV table during commit"
-                        );
-                        let on_error = std::mem::take(&mut *self.on_error.lock());
-                        let on_error_async = std::mem::take(&mut *self.on_error_async.lock());
-                        Self::execute_callbacks(on_error);
-                        Self::execute_async_callbacks(on_error_async).await;
-                        return Err(e.into());
-                    }
-                };
-
-                for (key, value) in pending.iter() {
-                    match value {
-                        Some(v) => {
-                            if let Err(e) = table.insert(key.as_slice(), v.as_slice()) {
-                                tracing::error!(
-                                    error = %e,
-                                    key_len = key.len(),
-                                    value_len = v.len(),
-                                    "Failed to insert key during commit - transaction will be rolled back"
-                                );
-                                // Note: write_txn is automatically rolled back when dropped (redb guarantee)
-                                tracing::debug!(
-                                    pending_changes = pending.len(),
-                                    "Write transaction aborted - no partial writes persisted"
-                                );
-                                let on_error = std::mem::take(&mut *self.on_error.lock());
-                                let on_error_async =
-                                    std::mem::take(&mut *self.on_error_async.lock());
-                                Self::execute_callbacks(on_error);
-                                Self::execute_async_callbacks(on_error_async).await;
-                                return Err(e.into());
-                            }
-                        }
-                        None => {
-                            if let Err(e) = table.remove(key.as_slice()) {
-                                tracing::error!(
-                                    error = %e,
-                                    key_len = key.len(),
-                                    "Failed to delete key during commit - transaction will be rolled back"
-                                );
-                                // Note: write_txn is automatically rolled back when dropped (redb guarantee)
-                                tracing::debug!(
-                                    pending_changes = pending.len(),
-                                    "Write transaction aborted - no partial writes persisted"
-                                );
-                                let on_error = std::mem::take(&mut *self.on_error.lock());
-                                let on_error_async =
-                                    std::mem::take(&mut *self.on_error_async.lock());
-                                Self::execute_callbacks(on_error);
-                                Self::execute_async_callbacks(on_error_async).await;
-                                return Err(e.into());
-                            }
-                        }
-                    }
-                }
-            }
-
-            if let Err(e) = write_txn.commit() {
-                tracing::error!(
-                    error = %e,
-                    pending_changes = pending.len(),
-                    "Failed to finalize commit - all changes rolled back"
-                );
-                // Note: redb guarantees atomicity - if commit fails, no changes are persisted
-                tracing::debug!("Commit failed at finalization stage - database state unchanged");
-                let on_error = std::mem::take(&mut *self.on_error.lock());
-                let on_error_async = std::mem::take(&mut *self.on_error_async.lock());
-                Self::execute_callbacks(on_error);
-                Self::execute_async_callbacks(on_error_async).await;
-                return Err(e.into());
-            }
-        }
-
-        // Mark as committed AFTER successful database commit
-        // This ensures the flag accurately reflects the transaction state
-        *self.committed.lock() = true;
-
-        // Execute success callbacks
-        let on_success = std::mem::take(&mut *self.on_success.lock());
-        let on_success_async = std::mem::take(&mut *self.on_success_async.lock());
-        Self::execute_callbacks(on_success);
-        Self::execute_async_callbacks(on_success_async).await;
-
-        Ok(())
-    }
-
-    fn discard(self: Box<Self>) {
-        // Note: active_txn_count is decremented by Drop impl when self is dropped
-        // at the end of this function.
-
-        *self.discarded.lock() = true;
-
-        // Execute sync discard callbacks
-        let on_discard = std::mem::take(&mut *self.on_discard.lock());
-        Self::execute_callbacks(on_discard);
-
-        // Handle async callbacks: spawn them in background with warning
-        let on_discard_async = std::mem::take(&mut *self.on_discard_async.lock());
-        if !on_discard_async.is_empty() {
-            let callback_count = on_discard_async.len();
-            tracing::warn!(
-                count = callback_count,
-                "Transaction has async discard callbacks. Spawning in background - they may not complete if process exits. Consider using commit() instead of discard() when async callbacks are registered."
-            );
-
-            tokio::spawn(async move {
-                Self::execute_async_callbacks(on_discard_async).await;
-                tracing::debug!(count = callback_count, "Async discard callbacks completed");
-            });
-        }
-    }
-
-    fn on_success(&mut self, callback: TxnCallback) {
-        self.on_success.lock().push(callback);
-    }
-
-    fn on_success_async(&mut self, callback: AsyncTxnCallback) {
-        self.on_success_async.lock().push(callback);
-    }
-
-    fn on_error(&mut self, callback: TxnCallback) {
-        self.on_error.lock().push(callback);
-    }
-
-    fn on_error_async(&mut self, callback: AsyncTxnCallback) {
-        self.on_error_async.lock().push(callback);
-    }
-
-    fn on_discard(&mut self, callback: TxnCallback) {
-        self.on_discard.lock().push(callback);
-    }
-
-    fn on_discard_async(&mut self, callback: AsyncTxnCallback) {
-        self.on_discard_async.lock().push(callback);
-    }
-
-    fn as_any(&self) -> &dyn std::any::Any {
-        self
-    }
-
-    fn as_any_mut(&mut self) -> &mut dyn std::any::Any {
-        self
-    }
-
-    fn is_readonly(&self) -> bool {
-        self.readonly
-    }
-
-    fn callback_count(&self) -> usize {
-        self.callback_counts().total()
-    }
-}
-
-/// Merging iterator that combines pre-materialized snapshot and pending changes.
-///
-/// Snapshot and pending items matching the query are materialized into Vecs at
-/// iterator creation. The merge itself happens on-demand during iteration via
-/// `next_merged()`. For large result sets, memory usage scales with the number
-/// of matching keys in the queried range.
-struct MergingIterator {
-    /// Items from snapshot (sorted ascending)
-    snapshot_items: Vec<(Vec<u8>, Vec<u8>)>,
-    /// Current position in snapshot
-    snapshot_pos: usize,
-
-    /// Pending changes (sorted ascending, None = deletion)
-    pending_items: Vec<(Vec<u8>, Option<Vec<u8>>)>,
-    /// Current position in pending
-    pending_pos: usize,
-
-    /// Whether iteration is reversed
-    reverse: bool,
-    /// Whether to return only keys
-    keys_only: bool,
-    /// Whether the iterator is closed
-    closed: bool,
-}
-
-impl MergingIterator {
-    fn new(
-        mut snapshot_items: Vec<(Vec<u8>, Vec<u8>)>,
-        mut pending_items: Vec<(Vec<u8>, Option<Vec<u8>>)>,
-        opts: IterOptions,
-    ) -> Self {
-        let reverse = opts.reverse();
-        if reverse {
-            snapshot_items.reverse();
-            pending_items.reverse();
-        }
-
-        Self {
-            snapshot_items,
-            snapshot_pos: 0,
-            pending_items,
-            pending_pos: 0,
-            reverse,
-            keys_only: opts.keys_only(),
-            closed: false,
-        }
-    }
-
-    /// Get the next merged key-value pair, handling overrides and deletions.
-    fn next_merged(&mut self) -> Option<(Vec<u8>, Vec<u8>)> {
-        loop {
-            let snap_key = self.snapshot_items.get(self.snapshot_pos).map(|(k, _)| k);
-            let pend_key = self.pending_items.get(self.pending_pos).map(|(k, _)| k);
-
-            match (snap_key, pend_key) {
-                (None, None) => return None,
-
-                (Some(_), None) => {
-                    let (key, value) = self.snapshot_items[self.snapshot_pos].clone();
-                    self.snapshot_pos += 1;
-                    return Some((key, value));
-                }
-
-                (None, Some(_)) => {
-                    let (key, value_opt) = self.pending_items[self.pending_pos].clone();
-                    self.pending_pos += 1;
-                    match value_opt {
-                        Some(value) => return Some((key, value)),
-                        None => continue, // Deletion of non-existent key, skip
-                    }
-                }
-
-                (Some(sk), Some(pk)) => {
-                    let cmp = if self.reverse {
-                        pk.cmp(sk) // Reversed: larger keys come first
-                    } else {
-                        sk.cmp(pk)
-                    };
-
-                    match cmp {
-                        std::cmp::Ordering::Less => {
-                            // Snapshot key comes first (no pending override)
-                            let (key, value) = self.snapshot_items[self.snapshot_pos].clone();
-                            self.snapshot_pos += 1;
-                            return Some((key, value));
-                        }
-                        std::cmp::Ordering::Greater => {
-                            // Pending key comes first (new key not in snapshot)
-                            let (key, value_opt) = self.pending_items[self.pending_pos].clone();
-                            self.pending_pos += 1;
-                            match value_opt {
-                                Some(value) => return Some((key, value)),
-                                None => continue, // Deletion of non-existent key
-                            }
-                        }
-                        std::cmp::Ordering::Equal => {
-                            // Same key: pending overrides snapshot
-                            let (key, value_opt) = self.pending_items[self.pending_pos].clone();
-                            self.snapshot_pos += 1; // Skip snapshot version
-                            self.pending_pos += 1;
-                            match value_opt {
-                                Some(value) => return Some((key, value)),
-                                None => continue, // Deletion
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    /// Binary search for seek position in a sorted Vec.
-    fn binary_search_position<T>(items: &[(Vec<u8>, T)], key: &[u8], reverse: bool) -> usize {
-        if reverse {
-            // Reversed: items are [k4, k3, k2, k1], find first <= key
-            items.partition_point(|(k, _)| k.as_slice() > key)
-        } else {
-            // Forward: items are [k1, k2, k3, k4], find first >= key
-            items.partition_point(|(k, _)| k.as_slice() < key)
-        }
-    }
-}
-
-#[async_trait]
-impl Iterator for MergingIterator {
-    async fn next(&mut self) -> Result<Option<KvPair>> {
-        if self.closed {
-            return Err(Error::Iterator("Iterator has been closed".into()));
-        }
-
-        match self.next_merged() {
-            Some((key, value)) => {
-                if self.keys_only {
-                    Ok(Some(KvPair::key_only(key)))
-                } else {
-                    Ok(Some(KvPair::new(key, value)))
-                }
-            }
-            None => Ok(None),
-        }
-    }
-
-    async fn close(&mut self) -> Result<()> {
-        self.closed = true;
-        Ok(())
-    }
-
-    async fn seek(&mut self, key: &[u8]) -> Result<bool> {
-        if self.closed {
-            return Err(Error::Iterator("Iterator has been closed".into()));
-        }
-
-        // Seek both iterators to the target position
-        self.snapshot_pos = Self::binary_search_position(&self.snapshot_items, key, self.reverse);
-        self.pending_pos = Self::binary_search_position(&self.pending_items, key, self.reverse);
-
-        // Check if there's actual visible data at or after the seek position.
-        // This accounts for pending deletions that might mask snapshot data.
-        // We do this by peeking at next_merged() without advancing the iterator.
-        let saved_snapshot_pos = self.snapshot_pos;
-        let saved_pending_pos = self.pending_pos;
-        let has_data = self.next_merged().is_some();
-        self.snapshot_pos = saved_snapshot_pos;
-        self.pending_pos = saved_pending_pos;
-
-        Ok(has_data)
-    }
-
-    async fn reset(&mut self) -> Result<()> {
-        if self.closed {
-            return Err(Error::Iterator("Iterator has been closed".into()));
-        }
-
-        self.snapshot_pos = 0;
-        self.pending_pos = 0;
-        Ok(())
-    }
-
-    fn is_valid(&self) -> bool {
-        !self.closed
-    }
-}
-
-// Convert redb errors to our error type with context and classification
-impl From<redb::Error> for Error {
-    fn from(err: redb::Error) -> Self {
-        match err {
-            // I/O errors
-            redb::Error::Io(io_err) => Error::Io(io_err.to_string()),
-            redb::Error::PreviousIo => {
-                tracing::error!("Previous I/O error - database must be closed and reopened");
-                Error::Io(
-                    "previous I/O error occurred - database must be closed and reopened".into(),
-                )
-            }
-
-            // Critical: Database corruption
-            redb::Error::Corrupted(ref msg) => {
-                tracing::error!(message = %msg, "Database corruption detected");
-                Error::Backend(format!(
-                    "database corrupted: {}. Recovery options: \
-                     (1) restore from backup, \
-                     (2) run check_integrity() to assess damage extent, \
-                     (3) delete database and resync from network peers (if available), \
-                     (4) check disk for hardware errors. \
-                     Consider preserving the corrupted file for forensic analysis before deletion.",
-                    msg
-                ))
-            }
-
-            // Lock poisoned: A thread panicked while holding a lock (fatal condition)
-            redb::Error::LockPoisoned(location) => {
-                tracing::error!(location = %location, "Lock poisoned - a thread panicked while holding a lock");
-                Error::Backend(format!(
-                    "internal error: lock poisoned at {} - database may be in undefined state",
-                    location
-                ))
-            }
-
-            // Database already open (useful for diagnosing lock issues)
-            redb::Error::DatabaseAlreadyOpen => {
-                tracing::warn!("Database is locked by another process");
-                Error::Backend(
-                    "database is locked by another process. \
-                     Check for other running processes or stale lock files"
-                        .into(),
-                )
-            }
-
-            // Upgrade required (file format migration needed)
-            redb::Error::UpgradeRequired(version) => {
-                tracing::warn!(version = version, "Database file format upgrade required");
-                Error::Backend(format!(
-                    "database uses file format version {} which requires upgrade. \
-                     Backup database and use redb migration tools",
-                    version
-                ))
-            }
-
-            // Transaction still in use (resource management issue, not a conflict)
-            redb::Error::ReadTransactionStillInUse(_) => {
-                tracing::warn!("Transaction still held by table or iterator");
-                Error::Backend(
-                    "transaction still in use - ensure all tables and iterators are dropped \
-                     before committing or discarding the transaction"
-                        .into(),
-                )
-            }
-
-            // Table errors with useful context
-            redb::Error::TableDoesNotExist(ref name) => {
-                Error::Backend(format!("table '{}' does not exist", name))
-            }
-            redb::Error::TableTypeMismatch { ref table, .. } => {
-                tracing::error!(table = %table, "Table type mismatch - possible schema corruption");
-                Error::Backend(format!("table type mismatch for '{}': {}", table, err))
-            }
-            redb::Error::TableAlreadyOpen(ref name, location) => {
-                tracing::warn!(table = %name, location = %location, "Table already open");
-                Error::Backend(format!("table '{}' already open at {}", name, location))
-            }
-
-            // Value size limit
-            redb::Error::ValueTooLarge(size) => {
-                const MAX_VALUE_SIZE: usize = 3 * 1024 * 1024 * 1024; // 3 GiB
-                Error::Backend(format!(
-                    "value too large: {} bytes exceeds redb maximum of {} bytes (3 GiB). \
-                     Consider chunking large values or using a different storage backend",
-                    size, MAX_VALUE_SIZE
-                ))
-            }
-
-            // Handle remaining variants (non-exhaustive enum)
-            other => Error::Backend(format!("redb error: {}", other)),
-        }
-    }
-}
-
-impl From<redb::DatabaseError> for Error {
-    fn from(err: redb::DatabaseError) -> Self {
-        match err {
-            redb::DatabaseError::DatabaseAlreadyOpen => {
-                tracing::warn!("Database is locked by another process");
-                Error::Backend(
-                    "database is locked by another process. \
-                     Check for other running processes or stale lock files"
-                        .into(),
-                )
-            }
-            redb::DatabaseError::UpgradeRequired(version) => {
-                tracing::warn!(version = version, "Database file format upgrade required");
-                Error::Backend(format!(
-                    "database uses file format version {} which requires upgrade. \
-                     Backup database and use redb migration tools",
-                    version
-                ))
-            }
-            redb::DatabaseError::RepairAborted => {
-                tracing::warn!("Database repair was aborted");
-                Error::Backend(
-                    "database repair was aborted before completion. \
-                     Database may be in inconsistent state - restore from backup recommended"
-                        .into(),
-                )
-            }
-            redb::DatabaseError::Storage(storage_err) => storage_err.into(),
-            // Handle future variants (non-exhaustive enum)
-            other => Error::Backend(format!("redb database error: {}", other)),
-        }
-    }
-}
-
-impl From<redb::TransactionError> for Error {
-    fn from(err: redb::TransactionError) -> Self {
-        match err {
-            // Resource management issue, NOT a transaction conflict
-            // This means a transaction is still held by a table or iterator
-            redb::TransactionError::ReadTransactionStillInUse(_) => {
-                tracing::warn!("Transaction still held by table or iterator");
-                Error::Backend(
-                    "transaction still in use - ensure all tables and iterators are dropped".into(),
-                )
-            }
-            redb::TransactionError::Storage(storage_err) => storage_err.into(),
-            // Handle future variants (non-exhaustive enum)
-            other => Error::Backend(format!("redb transaction error: {}", other)),
-        }
-    }
-}
-
-impl From<redb::TableError> for Error {
-    fn from(err: redb::TableError) -> Self {
-        match err {
-            redb::TableError::Storage(storage_err) => storage_err.into(),
-            redb::TableError::TableDoesNotExist(ref name) => {
-                Error::Backend(format!("table '{}' does not exist", name))
-            }
-            redb::TableError::TableTypeMismatch { ref table, .. } => {
-                tracing::error!(table = %table, "Table type mismatch - possible schema corruption");
-                Error::Backend(format!("table type mismatch for '{}': {}", table, err))
-            }
-            redb::TableError::TableAlreadyOpen(ref name, location) => {
-                tracing::warn!(table = %name, location = %location, "Table already open");
-                Error::Backend(format!("table '{}' already open at {}", name, location))
-            }
-            redb::TableError::TableIsMultimap(ref name) => {
-                Error::Backend(format!("table '{}' is a multimap table", name))
-            }
-            redb::TableError::TableIsNotMultimap(ref name) => {
-                Error::Backend(format!("table '{}' is not a multimap table", name))
-            }
-            redb::TableError::TableExists(ref name) => {
-                Error::Backend(format!("table '{}' already exists", name))
-            }
-            // Handle future variants (non-exhaustive enum)
-            other => Error::Backend(format!("redb table error: {}", other)),
-        }
-    }
-}
-
-impl From<redb::StorageError> for Error {
-    fn from(err: redb::StorageError) -> Self {
-        match err {
-            redb::StorageError::Io(io_err) => Error::Io(io_err.to_string()),
-            redb::StorageError::PreviousIo => {
-                tracing::error!("Previous I/O error - database must be closed and reopened");
-                Error::Io(
-                    "previous I/O error occurred - database must be closed and reopened".into(),
-                )
-            }
-            redb::StorageError::Corrupted(ref msg) => {
-                tracing::error!(message = %msg, "Database corruption detected");
-                Error::Backend(format!(
-                    "database corrupted: {}. Recovery options: \
-                     (1) restore from backup, \
-                     (2) run check_integrity() to assess damage extent, \
-                     (3) delete database and resync from network peers (if available), \
-                     (4) check disk for hardware errors. \
-                     Consider preserving the corrupted file for forensic analysis before deletion.",
-                    msg
-                ))
-            }
-            redb::StorageError::ValueTooLarge(size) => {
-                // redb has a 3GB maximum value size
-                const MAX_VALUE_SIZE: usize = 3 * 1024 * 1024 * 1024; // 3 GiB
-                Error::Backend(format!(
-                    "value too large: {} bytes exceeds redb maximum of {} bytes (3 GiB). \
-                     Consider chunking large values or using a different storage backend",
-                    size, MAX_VALUE_SIZE
-                ))
-            }
-            // CRITICAL: LockPoisoned is NOT a transaction conflict!
-            // It indicates a thread panicked while holding a lock - this is fatal.
-            redb::StorageError::LockPoisoned(location) => {
-                tracing::error!(location = %location, "Lock poisoned - a thread panicked while holding a lock");
-                Error::Backend(format!(
-                    "internal error: lock poisoned at {} - database may be in undefined state",
-                    location
-                ))
-            }
-            // Handle future variants (non-exhaustive enum)
-            other => Error::Backend(format!("redb storage error: {}", other)),
-        }
-    }
-}
-
-impl From<redb::CommitError> for Error {
-    fn from(err: redb::CommitError) -> Self {
-        match err {
-            // Delegate storage errors to the StorageError handler for consistent classification
-            redb::CommitError::Storage(storage_err) => storage_err.into(),
-            // Handle future variants (non-exhaustive enum)
-            other => Error::Backend(format!("redb commit error: {}", other)),
-        }
-    }
-}
-
-// ============================================================================
-// SHARED TEST SUITE - Run same tests against all backends
-// ============================================================================
-#[cfg(test)]
-mod shared_tests {
-    use super::*;
-    use crate::corekv::Dropable;
-    use crate::generate_backend_concurrency_tests;
-    use crate::generate_backend_dropable_tests;
-    use crate::generate_backend_tests;
-    use tempfile::TempDir;
-
-    /// Test wrapper that holds both store and temp directory for cleanup.
-    /// When this wrapper is dropped, the TempDir is automatically cleaned up.
-    struct TestRedbStore {
-        store: RedbStore,
-        _temp_dir: TempDir,
-    }
-
-    #[async_trait]
-    impl Store for TestRedbStore {
-        async fn new_txn(&self, readonly: bool) -> Result<Box<dyn Txn>> {
-            self.store.new_txn(readonly).await
-        }
-        async fn close(&self) -> Result<()> {
-            self.store.close().await
-        }
-    }
-
-    #[async_trait]
-    impl Dropable for TestRedbStore {
-        async fn drop_all(&self) -> Result<()> {
-            self.store.drop_all().await
-        }
-    }
-
-    async fn create_store() -> TestRedbStore {
-        let temp_dir = TempDir::new().unwrap();
-        let path = temp_dir.path().join("test.redb");
-        let store = RedbStore::open(&path).unwrap();
-        TestRedbStore {
-            store,
-            _temp_dir: temp_dir,
-        }
-    }
-
-    async fn create_arc_store() -> Arc<TestRedbStore> {
-        Arc::new(create_store().await)
-    }
-
-    // Generate all standard backend tests
-    generate_backend_tests!(create_store);
-
-    // Generate concurrency tests
-    generate_backend_concurrency_tests!(create_arc_store);
-
-    // Generate Dropable tests (RedbStore implements Dropable)
-    generate_backend_dropable_tests!(create_store);
-}
-
-// ============================================================================
-// REDB-SPECIFIC TESTS - Persistence and specific behaviors
-// ============================================================================
-#[cfg(test)]
-mod redb_specific_tests {
-    use super::*;
-    use tempfile::TempDir;
-
-    #[tokio::test]
-    async fn test_redb_data_survives_close_reopen() {
-        let temp_dir = TempDir::new().unwrap();
-        let path = temp_dir.path().join("test.redb");
-
-        // Write data and close
-        {
-            let store = RedbStore::open(&path).unwrap();
-            let mut txn = store.new_txn(false).await.unwrap();
-            txn.set(b"persistent_key", b"persistent_value")
-                .await
-                .unwrap();
-            txn.commit().await.unwrap();
-            store.close().await.unwrap();
-        }
-
-        // Reopen and verify
-        {
-            let store = RedbStore::open(&path).unwrap();
-            let txn = store.new_txn(true).await.unwrap();
-            assert_eq!(
-                txn.get(b"persistent_key").await.unwrap(),
-                Some(b"persistent_value".to_vec()),
-                "Data should survive close/reopen"
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn test_redb_uncommitted_data_lost_on_reopen() {
-        let temp_dir = TempDir::new().unwrap();
-        let path = temp_dir.path().join("test.redb");
-
-        // Write data but DON'T commit
-        {
-            let store = RedbStore::open(&path).unwrap();
-            let mut txn = store.new_txn(false).await.unwrap();
-            txn.set(b"uncommitted_key", b"value").await.unwrap();
-            // No commit! Discard.
-            txn.discard();
-            store.close().await.unwrap();
-        }
-
-        // Reopen - uncommitted data should be gone
-        {
-            let store = RedbStore::open(&path).unwrap();
-            let txn = store.new_txn(true).await.unwrap();
-            assert_eq!(
-                txn.get(b"uncommitted_key").await.unwrap(),
-                None,
-                "Uncommitted data should not survive reopen"
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn test_redb_persistence_through_multiple_sessions() {
-        let temp_dir = TempDir::new().unwrap();
-        let path = temp_dir.path().join("test.redb");
-
-        // Session 1: Write keys
-        {
-            let store = RedbStore::open(&path).unwrap();
-            let mut txn = store.new_txn(false).await.unwrap();
-            txn.set(b"key1", b"value1").await.unwrap();
-            txn.set(b"key2", b"value2").await.unwrap();
-            txn.commit().await.unwrap();
-        }
-
-        // Session 2: Modify and add
-        {
-            let store = RedbStore::open(&path).unwrap();
-            let mut txn = store.new_txn(false).await.unwrap();
-            txn.set(b"key1", b"modified").await.unwrap();
-            txn.set(b"key3", b"value3").await.unwrap();
-            txn.delete(b"key2").await.unwrap();
-            txn.commit().await.unwrap();
-        }
-
-        // Session 3: Verify all changes
-        {
-            let store = RedbStore::open(&path).unwrap();
-            let txn = store.new_txn(true).await.unwrap();
-            assert_eq!(txn.get(b"key1").await.unwrap(), Some(b"modified".to_vec()));
-            assert_eq!(txn.get(b"key2").await.unwrap(), None);
-            assert_eq!(txn.get(b"key3").await.unwrap(), Some(b"value3".to_vec()));
-        }
-    }
-
-    #[tokio::test]
-    async fn test_redb_snapshot_isolation() {
-        let temp_dir = TempDir::new().unwrap();
-        let store = Arc::new(RedbStore::open(temp_dir.path().join("test.redb")).unwrap());
-
-        // Setup initial value
-        let mut setup = store.new_txn(false).await.unwrap();
-        setup.set(b"key", b"initial").await.unwrap();
-        setup.commit().await.unwrap();
-
-        // Start a reader BEFORE the write
-        let reader = store.new_txn(true).await.unwrap();
-
-        // Concurrent writer commits
-        let mut writer = store.new_txn(false).await.unwrap();
-        writer.set(b"key", b"modified").await.unwrap();
-        writer.commit().await.unwrap();
-
-        // Reader should see the ORIGINAL value (snapshot isolation)
-        let value = reader.get(b"key").await.unwrap();
-        assert_eq!(
-            value,
-            Some(b"initial".to_vec()),
-            "Reader should see original value (snapshot isolation)"
-        );
-
-        // A new reader should see the modified value
-        let new_reader = store.new_txn(true).await.unwrap();
-        let new_value = new_reader.get(b"key").await.unwrap();
-        assert_eq!(
-            new_value,
-            Some(b"modified".to_vec()),
-            "New reader should see committed changes"
-        );
-    }
-
-    // Note: Tests for "operations after discard/commit" are unnecessary because
-    // Rust's ownership system enforces this at compile time - discard() and commit()
-    // take `self: Box<Self>`, consuming the transaction and preventing further use.
-
-    #[tokio::test]
-    async fn test_redb_active_transaction_tracking() {
-        let temp_dir = TempDir::new().unwrap();
-        let store = RedbStore::open(temp_dir.path().join("test.redb")).unwrap();
-
-        assert_eq!(
-            store.active_transaction_count(),
-            0,
-            "No active transactions initially"
-        );
-
-        // Create a transaction
-        let txn1 = store.new_txn(true).await.unwrap();
-        assert_eq!(
-            store.active_transaction_count(),
-            1,
-            "One active transaction"
-        );
-
-        // Create another
-        let txn2 = store.new_txn(false).await.unwrap();
-        assert_eq!(
-            store.active_transaction_count(),
-            2,
-            "Two active transactions"
-        );
-
-        // Discard one
-        txn1.discard();
-        assert_eq!(
-            store.active_transaction_count(),
-            1,
-            "One active after discard"
-        );
-
-        // Commit the other
-        txn2.commit().await.unwrap();
-        assert_eq!(
-            store.active_transaction_count(),
-            0,
-            "None active after commit"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_redb_close_waits_for_transactions() {
-        let temp_dir = TempDir::new().unwrap();
-        let store = Arc::new(RedbStore::open(temp_dir.path().join("test.redb")).unwrap());
-
-        // Create a transaction
-        let txn = store.new_txn(true).await.unwrap();
-        assert_eq!(store.active_transaction_count(), 1);
-
-        // Spawn close in background
-        let store_clone = Arc::clone(&store);
-        let close_handle = tokio::spawn(async move {
-            store_clone.close().await.unwrap();
-        });
-
-        // Small delay to let close start
-        tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-
-        // Discard the transaction
-        txn.discard();
-
-        // Close should complete
-        tokio::time::timeout(std::time::Duration::from_secs(2), close_handle)
-            .await
-            .expect("Close should complete")
-            .expect("Close should succeed");
-
-        assert_eq!(store.active_transaction_count(), 0);
-    }
-
-    #[tokio::test]
-    async fn test_redb_operations_on_closed_store_fail() {
-        let temp_dir = TempDir::new().unwrap();
-        let store = RedbStore::open(temp_dir.path().join("test.redb")).unwrap();
-
-        store.close().await.unwrap();
-
-        // New transactions should fail
-        let result = store.new_txn(true).await;
-        assert!(result.is_err(), "new_txn should fail on closed store");
-    }
-
-    #[tokio::test]
-    async fn test_redb_iterator_prefix_filtering() {
-        let temp_dir = TempDir::new().unwrap();
-        let store = RedbStore::open(temp_dir.path().join("test.redb")).unwrap();
-
-        // Insert keys with different prefixes
-        let mut txn = store.new_txn(false).await.unwrap();
-        txn.set(b"user:1", b"alice").await.unwrap();
-        txn.set(b"user:2", b"bob").await.unwrap();
-        txn.set(b"user:3", b"carol").await.unwrap();
-        txn.set(b"doc:1", b"document1").await.unwrap();
-        txn.set(b"doc:2", b"document2").await.unwrap();
-        txn.commit().await.unwrap();
-
-        // Test prefix iteration
-        let txn = store.new_txn(true).await.unwrap();
-        let opts = IterOptions::new().with_prefix(b"user:".to_vec());
-        let mut iter = txn.iterator(opts).await.unwrap();
-
-        let mut count = 0;
-        while let Some(kv) = iter.next().await.unwrap() {
-            assert!(kv.key.starts_with(b"user:"), "Key should have prefix");
-            count += 1;
-        }
-        assert_eq!(count, 3, "Should have 3 user keys");
-
-        // Test doc prefix
-        let opts = IterOptions::new().with_prefix(b"doc:".to_vec());
-        let mut iter = txn.iterator(opts).await.unwrap();
-
-        let mut count = 0;
-        while let Some(kv) = iter.next().await.unwrap() {
-            assert!(kv.key.starts_with(b"doc:"), "Key should have prefix");
-            count += 1;
-        }
-        assert_eq!(count, 2, "Should have 2 doc keys");
-    }
-
-    #[tokio::test]
-    async fn test_redb_iterator_range_filtering() {
-        let temp_dir = TempDir::new().unwrap();
-        let store = RedbStore::open(temp_dir.path().join("test.redb")).unwrap();
-
-        // Insert alphabetically ordered keys
-        let mut txn = store.new_txn(false).await.unwrap();
-        txn.set(b"a", b"1").await.unwrap();
-        txn.set(b"b", b"2").await.unwrap();
-        txn.set(b"c", b"3").await.unwrap();
-        txn.set(b"d", b"4").await.unwrap();
-        txn.set(b"e", b"5").await.unwrap();
-        txn.commit().await.unwrap();
-
-        // Test range iteration [b, d)
-        let txn = store.new_txn(true).await.unwrap();
-        let opts = IterOptions::new()
-            .with_start(b"b".to_vec())
-            .with_end(b"d".to_vec());
-        let mut iter = txn.iterator(opts).await.unwrap();
-
-        let keys: Vec<_> = {
-            let mut keys = vec![];
-            while let Some(kv) = iter.next().await.unwrap() {
-                keys.push(kv.key);
-            }
-            keys
-        };
-
-        assert_eq!(keys.len(), 2, "Should have keys b and c");
-        assert_eq!(keys[0], b"b");
-        assert_eq!(keys[1], b"c");
-    }
-
-    #[tokio::test]
-    async fn test_redb_iterator_reverse() {
-        let temp_dir = TempDir::new().unwrap();
-        let store = RedbStore::open(temp_dir.path().join("test.redb")).unwrap();
-
-        let mut txn = store.new_txn(false).await.unwrap();
-        txn.set(b"a", b"1").await.unwrap();
-        txn.set(b"b", b"2").await.unwrap();
-        txn.set(b"c", b"3").await.unwrap();
-        txn.commit().await.unwrap();
-
-        let txn = store.new_txn(true).await.unwrap();
-        let opts = IterOptions::new().with_reverse(true);
-        let mut iter = txn.iterator(opts).await.unwrap();
-
-        let mut keys = vec![];
-        while let Some(kv) = iter.next().await.unwrap() {
-            keys.push(kv.key);
-        }
-
-        assert_eq!(keys, vec![b"c".to_vec(), b"b".to_vec(), b"a".to_vec()]);
-    }
-
-    #[tokio::test]
-    async fn test_redb_empty_key_rejected() {
-        let temp_dir = TempDir::new().unwrap();
-        let store = RedbStore::open(temp_dir.path().join("test.redb")).unwrap();
-
-        let mut txn = store.new_txn(false).await.unwrap();
-
-        // Empty key should be rejected
-        let result = txn.set(b"", b"value").await;
-        assert!(result.is_err(), "Empty key should be rejected");
-
-        let result = txn.get(b"").await;
-        assert!(result.is_err(), "Empty key get should be rejected");
-
-        let result = txn.delete(b"").await;
-        assert!(result.is_err(), "Empty key delete should be rejected");
-    }
-
-    #[tokio::test]
-    async fn test_redb_read_only_txn_rejects_writes() {
-        let temp_dir = TempDir::new().unwrap();
-        let store = RedbStore::open(temp_dir.path().join("test.redb")).unwrap();
-
-        let mut txn = store.new_txn(true).await.unwrap(); // read-only
-
-        let result = txn.set(b"key", b"value").await;
-        assert!(result.is_err(), "Read-only txn should reject set");
-
-        let result = txn.delete(b"key").await;
-        assert!(result.is_err(), "Read-only txn should reject delete");
-    }
-
-    #[tokio::test]
-    async fn test_redb_pending_changes_merged_with_snapshot() {
-        let temp_dir = TempDir::new().unwrap();
-        let store = RedbStore::open(temp_dir.path().join("test.redb")).unwrap();
-
-        // Setup initial data
-        let mut txn = store.new_txn(false).await.unwrap();
-        txn.set(b"existing", b"original").await.unwrap();
-        txn.commit().await.unwrap();
-
-        // Start a new transaction with pending changes
-        let mut txn = store.new_txn(false).await.unwrap();
-        txn.set(b"existing", b"modified").await.unwrap();
-        txn.set(b"new_key", b"new_value").await.unwrap();
-
-        // Pending changes should be visible
-        assert_eq!(
-            txn.get(b"existing").await.unwrap(),
-            Some(b"modified".to_vec()),
-            "Should see pending modification"
-        );
-        assert_eq!(
-            txn.get(b"new_key").await.unwrap(),
-            Some(b"new_value".to_vec()),
-            "Should see pending new key"
-        );
-
-        // Iterator should also merge pending changes
-        let opts = IterOptions::new();
-        let mut iter = txn.iterator(opts).await.unwrap();
-        let mut found_keys = std::collections::HashSet::new();
-        while let Some(kv) = iter.next().await.unwrap() {
-            found_keys.insert(kv.key);
-        }
-        assert!(found_keys.contains(&b"existing".to_vec()));
-        assert!(found_keys.contains(&b"new_key".to_vec()));
-    }
-
-    #[tokio::test]
-    async fn test_redb_pending_delete_removes_from_snapshot() {
-        let temp_dir = TempDir::new().unwrap();
-        let store = RedbStore::open(temp_dir.path().join("test.redb")).unwrap();
-
-        // Setup initial data
-        let mut txn = store.new_txn(false).await.unwrap();
-        txn.set(b"to_delete", b"value").await.unwrap();
-        txn.set(b"to_keep", b"value").await.unwrap();
-        txn.commit().await.unwrap();
-
-        // Delete in a new transaction
-        let mut txn = store.new_txn(false).await.unwrap();
-        txn.delete(b"to_delete").await.unwrap();
-
-        // Deleted key should not be visible
-        assert_eq!(
-            txn.get(b"to_delete").await.unwrap(),
-            None,
-            "Deleted key should not be visible"
-        );
-        assert!(!txn.has(b"to_delete").await.unwrap());
-
-        // Iterator should not include deleted key
-        let opts = IterOptions::new();
-        let mut iter = txn.iterator(opts).await.unwrap();
-        while let Some(kv) = iter.next().await.unwrap() {
-            assert_ne!(
-                kv.key, b"to_delete",
-                "Deleted key should not appear in iterator"
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn test_redb_directory_handling() {
-        let temp_dir = TempDir::new().unwrap();
-        let dir_path = temp_dir.path().to_path_buf();
-
-        // Opening with directory path should work (creates data.redb inside)
-        {
-            let store = RedbStore::open(&dir_path).unwrap();
-            let mut txn = store.new_txn(false).await.unwrap();
-            txn.set(b"key", b"value").await.unwrap();
-            txn.commit().await.unwrap();
-            store.close().await.unwrap();
-            // Store dropped here, releasing the lock
-        }
-
-        // Verify the database file was created inside the directory
-        let db_path = dir_path.join("data.redb");
-        assert!(db_path.exists(), "data.redb should be created in directory");
-
-        // Reopen and verify data
-        {
-            let store = RedbStore::open(&dir_path).unwrap();
-            let txn = store.new_txn(true).await.unwrap();
-            assert_eq!(txn.get(b"key").await.unwrap(), Some(b"value".to_vec()));
-        }
-    }
-
-    #[tokio::test]
-    async fn test_redb_large_value_handling() {
-        let temp_dir = TempDir::new().unwrap();
-        let path = temp_dir.path().join("test.redb");
-
-        // Test with 5MB value
-        let large_value = vec![0xABu8; 5 * 1024 * 1024];
-
-        // Write and verify retrieval
-        {
-            let store = RedbStore::open(&path).unwrap();
-
-            let mut txn = store.new_txn(false).await.unwrap();
-            txn.set(b"large_key", &large_value).await.unwrap();
-            txn.commit().await.unwrap();
-
-            // Verify retrieval
-            let txn = store.new_txn(true).await.unwrap();
-            let retrieved = txn.get(b"large_key").await.unwrap();
-            assert_eq!(
-                retrieved.as_ref().map(|v| v.len()),
-                Some(5 * 1024 * 1024),
-                "Large value should be retrievable"
-            );
-            assert_eq!(
-                retrieved.as_ref().map(|v| v[0]),
-                Some(0xAB),
-                "Large value content should match"
-            );
-
-            // Clean up transaction before closing
-            txn.discard();
-            store.close().await.unwrap();
-        }
-
-        // Verify persistence after reopen
-        {
-            let store = RedbStore::open(&path).unwrap();
-            let txn = store.new_txn(true).await.unwrap();
-            let retrieved = txn.get(b"large_key").await.unwrap();
-            assert_eq!(
-                retrieved.map(|v| v.len()),
-                Some(5 * 1024 * 1024),
-                "Large value should survive persistence"
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn test_redb_new_txn_rejected_during_close() {
-        let temp_dir = TempDir::new().unwrap();
-        let store = Arc::new(RedbStore::open(temp_dir.path().join("test.redb")).unwrap());
-
-        // Create a transaction to keep the store busy
-        let txn = store.new_txn(true).await.unwrap();
-        assert_eq!(store.active_transaction_count(), 1);
-
-        // Start close in background (will wait for active transactions)
-        let store_clone = Arc::clone(&store);
-        let close_handle = tokio::spawn(async move {
-            // Small delay to ensure close starts
-            tokio::time::sleep(std::time::Duration::from_millis(10)).await;
-            store_clone.close().await
-        });
-
-        // Wait for close to mark the store as closed
-        tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-
-        // New transaction should be rejected because store is closing
-        let result = store.new_txn(true).await;
-        assert!(result.is_err(), "new_txn should fail when store is closing");
-
-        // Clean up: discard the blocking transaction so close can complete
-        txn.discard();
-
-        // Wait for close to complete
-        let close_result = tokio::time::timeout(std::time::Duration::from_secs(2), close_handle)
-            .await
-            .expect("Close should complete")
-            .expect("Close task should not panic");
-
-        assert!(close_result.is_ok(), "Close should succeed");
-    }
-
-    #[tokio::test]
-    async fn test_redb_custom_cache_size() {
-        use super::super::redb_config::RedbStoreOptions;
-
-        let temp_dir = TempDir::new().unwrap();
-        let path = temp_dir.path().join("test.redb");
-
-        // Open with custom cache size (16MB)
-        let opts = RedbStoreOptions::new().with_cache_size(16 * 1024 * 1024);
-        let store = RedbStore::open_with_options(&path, opts).unwrap();
-
-        // Verify store works normally
-        let mut txn = store.new_txn(false).await.unwrap();
-        txn.set(b"key", b"value").await.unwrap();
-        txn.commit().await.unwrap();
-
-        let txn = store.new_txn(true).await.unwrap();
-        assert_eq!(
-            txn.get(b"key").await.unwrap(),
-            Some(b"value".to_vec()),
-            "Store with custom cache should work normally"
-        );
-        txn.discard();
-    }
-
-    #[tokio::test]
-    async fn test_redb_error_callback_on_discarded_commit() {
-        use std::sync::atomic::{AtomicBool, Ordering};
-
-        let temp_dir = TempDir::new().unwrap();
-        let store = RedbStore::open(temp_dir.path().join("test.redb")).unwrap();
-
-        let error_called = Arc::new(AtomicBool::new(false));
-        let success_called = Arc::new(AtomicBool::new(false));
-
-        let error_flag = Arc::clone(&error_called);
-        let success_flag = Arc::clone(&success_called);
-
-        let mut txn = store.new_txn(false).await.unwrap();
-        txn.set(b"key", b"value").await.unwrap();
-
-        // Register callbacks
-        txn.on_success(Box::new(move || {
-            success_flag.store(true, Ordering::SeqCst);
-        }));
-        txn.on_error(Box::new(move || {
-            error_flag.store(true, Ordering::SeqCst);
-        }));
-
-        // Discard the transaction first
-        txn.discard();
-
-        // Try to commit after discard - this should fail and call error callback
-        // Note: Since discard() consumes self, we can't actually call commit() after.
-        // However, we CAN test the error path by trying to commit a new discarded txn.
-        // The Rust ownership model prevents the actual scenario, but we can verify
-        // that error callbacks ARE called when commit fails due to other reasons.
-
-        // Instead, test that error callback IS invoked when commit on discarded txn happens
-        // by checking the callback mechanism works on normal success path
-        assert!(
-            !error_called.load(Ordering::SeqCst),
-            "Error callback should not be called on discard"
-        );
-        assert!(
-            !success_called.load(Ordering::SeqCst),
-            "Success callback should not be called on discard"
-        );
-
-        // Now test success callback works
-        let success_called2 = Arc::new(AtomicBool::new(false));
-        let success_flag2 = Arc::clone(&success_called2);
-
-        let mut txn = store.new_txn(false).await.unwrap();
-        txn.set(b"key2", b"value2").await.unwrap();
-        txn.on_success(Box::new(move || {
-            success_flag2.store(true, Ordering::SeqCst);
-        }));
-        txn.commit().await.unwrap();
-
-        assert!(
-            success_called2.load(Ordering::SeqCst),
-            "Success callback should be called on commit"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_redb_async_error_callback() {
-        use std::sync::atomic::{AtomicBool, Ordering};
-
-        let temp_dir = TempDir::new().unwrap();
-        let store = RedbStore::open(temp_dir.path().join("test.redb")).unwrap();
-
-        let async_success_called = Arc::new(AtomicBool::new(false));
-        let async_flag = Arc::clone(&async_success_called);
-
-        let mut txn = store.new_txn(false).await.unwrap();
-        txn.set(b"key", b"value").await.unwrap();
-
-        // Register async callback
-        txn.on_success_async(Box::new(move || {
-            let flag = async_flag;
-            Box::pin(async move {
-                flag.store(true, Ordering::SeqCst);
-            })
-        }));
-
-        txn.commit().await.unwrap();
-
-        assert!(
-            async_success_called.load(Ordering::SeqCst),
-            "Async success callback should be called and awaited on commit"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_redb_iterator_seek_on_empty_store() {
-        let temp_dir = TempDir::new().unwrap();
-        let store = RedbStore::open(temp_dir.path().join("test.redb")).unwrap();
-
-        let txn = store.new_txn(true).await.unwrap();
-        let opts = IterOptions::new();
-        let mut iter = txn.iterator(opts).await.unwrap();
-
-        // Seek on empty store should return false
-        let found = iter.seek(b"any_key").await.unwrap();
-        assert!(!found, "Seek on empty store should return false");
-
-        // Next should return None
-        let item = iter.next().await.unwrap();
-        assert!(item.is_none(), "Next on empty store should return None");
-
-        // Reset should succeed
-        iter.reset().await.unwrap();
-        assert!(
-            iter.is_valid(),
-            "Iterator should still be valid after reset"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_redb_many_keys_persistence() {
-        let temp_dir = TempDir::new().unwrap();
-        let path = temp_dir.path().join("test.redb");
-
-        // Write 1000 keys
-        {
-            let store = RedbStore::open(&path).unwrap();
-            let mut txn = store.new_txn(false).await.unwrap();
-            for i in 0..1000u32 {
-                let key = format!("key_{:05}", i);
-                let value = format!("value_{}", i);
-                txn.set(key.as_bytes(), value.as_bytes()).await.unwrap();
-            }
-            txn.commit().await.unwrap();
-            store.close().await.unwrap();
-        }
-
-        // Reopen and verify all keys
-        {
-            let store = RedbStore::open(&path).unwrap();
-            let txn = store.new_txn(true).await.unwrap();
-
-            // Verify specific keys
-            for i in [0, 100, 500, 999].iter() {
-                let key = format!("key_{:05}", i);
-                let expected = format!("value_{}", i);
-                let value = txn.get(key.as_bytes()).await.unwrap();
-                assert_eq!(
-                    value,
-                    Some(expected.into_bytes()),
-                    "Key {} should be retrievable after persistence",
-                    key
-                );
-            }
-
-            // Verify count via iterator
-            let opts = IterOptions::new();
-            let mut iter = txn.iterator(opts).await.unwrap();
-            let mut count = 0;
-            while iter.next().await.unwrap().is_some() {
-                count += 1;
-            }
-            assert_eq!(count, 1000, "Should have 1000 keys after persistence");
-        }
-    }
-
-    #[tokio::test]
-    async fn test_redb_drop_all_with_active_transactions() {
-        use crate::corekv::Dropable;
-
-        let temp_dir = TempDir::new().unwrap();
-        let store = Arc::new(RedbStore::open(temp_dir.path().join("test.redb")).unwrap());
-
-        // Setup initial data
-        {
-            let mut txn = store.new_txn(false).await.unwrap();
-            txn.set(b"key1", b"value1").await.unwrap();
-            txn.set(b"key2", b"value2").await.unwrap();
-            txn.commit().await.unwrap();
-        }
-
-        // Start a read transaction before drop_all
-        let reader = store.new_txn(true).await.unwrap();
-
-        // Verify reader can see the data
-        assert_eq!(
-            reader.get(b"key1").await.unwrap(),
-            Some(b"value1".to_vec()),
-            "Reader should see initial data"
-        );
-
-        // drop_all should succeed even with active readers
-        // (redb allows this - readers see their snapshot, drop_all creates new state)
-        store.drop_all().await.unwrap();
-
-        // Reader should still see the snapshot data (MVCC isolation)
-        assert_eq!(
-            reader.get(b"key1").await.unwrap(),
-            Some(b"value1".to_vec()),
-            "Reader should still see snapshot data after drop_all"
-        );
-
-        // A new transaction should see empty store
-        let new_reader = store.new_txn(true).await.unwrap();
-        assert_eq!(
-            new_reader.get(b"key1").await.unwrap(),
-            None,
-            "New reader should see empty store after drop_all"
-        );
-
-        // Clean up transactions before closing
-        reader.discard();
-        new_reader.discard();
-    }
-
-    #[tokio::test]
-    async fn test_redb_rapid_transaction_cycles() {
-        use std::sync::atomic::{AtomicUsize, Ordering};
-
-        let temp_dir = TempDir::new().unwrap();
-        let store = Arc::new(RedbStore::open(temp_dir.path().join("test.redb")).unwrap());
-
-        let completed = Arc::new(AtomicUsize::new(0));
-        let num_tasks = 50;
-        let cycles_per_task = 20;
-
-        let mut handles = vec![];
-
-        for task_id in 0..num_tasks {
-            let store = Arc::clone(&store);
-            let completed = Arc::clone(&completed);
-
-            handles.push(tokio::spawn(async move {
-                for cycle in 0..cycles_per_task {
-                    // Alternate between read-only and read-write transactions
-                    let readonly = cycle % 2 == 0;
-                    let txn = store.new_txn(readonly).await.unwrap();
-
-                    // Do some work
-                    if !readonly {
-                        let mut txn = txn;
-                        let key = format!("task_{}_cycle_{}", task_id, cycle);
-                        txn.set(key.as_bytes(), b"value").await.unwrap();
-
-                        // Alternate between commit and discard
-                        if cycle % 3 == 0 {
-                            txn.discard();
-                        } else {
-                            txn.commit().await.unwrap();
-                        }
-                    } else {
-                        // Read-only: just read and discard
-                        let _ = txn.has(b"some_key").await;
-                        txn.discard();
-                    }
-
-                    completed.fetch_add(1, Ordering::SeqCst);
-                }
-            }));
-        }
-
-        // Wait for all tasks to complete
-        for handle in handles {
-            handle.await.unwrap();
-        }
-
-        // Verify all cycles completed
-        assert_eq!(
-            completed.load(Ordering::SeqCst),
-            num_tasks * cycles_per_task,
-            "All transaction cycles should complete"
-        );
-
-        // Verify no transactions are leaked
-        assert_eq!(
-            store.active_transaction_count(),
-            0,
-            "No active transactions should remain after all cycles complete"
-        );
-
-        // Store should close cleanly without timeout
-        store.close().await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_redb_close_timeout_returns_error() {
-        let temp_dir = TempDir::new().unwrap();
-        let store = Arc::new(RedbStore::open(temp_dir.path().join("test.redb")).unwrap());
-
-        // Create a transaction that we intentionally don't close
-        let _held_txn = store.new_txn(true).await.unwrap();
-
-        assert_eq!(
-            store.active_transaction_count(),
-            1,
-            "Should have one active transaction"
-        );
-
-        // Close should timeout and return an error (timeout is 5 seconds)
-        // We use a shorter test by checking the error is returned
-        let start = std::time::Instant::now();
-        let result = store.close().await;
-
-        // Verify that close took approximately 5 seconds (with some tolerance)
-        let elapsed = start.elapsed();
-        assert!(
-            elapsed >= std::time::Duration::from_secs(4),
-            "Close should have waited at least 4 seconds, but took {:?}",
-            elapsed
-        );
-
-        // Verify error was returned
-        assert!(
-            result.is_err(),
-            "Close should return error when transactions are still active"
-        );
-
-        let err = result.unwrap_err();
-        let err_msg = err.to_string();
-        assert!(
-            err_msg.contains("Close timeout"),
-            "Error should mention timeout: {}",
-            err_msg
-        );
-        assert!(
-            err_msg.contains("still active"),
-            "Error should mention active transactions: {}",
-            err_msg
-        );
-    }
-
-    // =========================================================================
-    // HIGH-CONTENTION STRESS TESTS
-    // =========================================================================
-
-    #[tokio::test]
-    async fn test_redb_high_contention_100_concurrent_txns() {
-        use std::sync::atomic::{AtomicUsize, Ordering};
-
-        let temp_dir = TempDir::new().unwrap();
-        let store = Arc::new(RedbStore::open(temp_dir.path().join("test.redb")).unwrap());
-
-        let completed = Arc::new(AtomicUsize::new(0));
-        let num_tasks = 100;
-
-        let mut handles = vec![];
-
-        for i in 0..num_tasks {
-            let store = Arc::clone(&store);
-            let completed = Arc::clone(&completed);
-
-            handles.push(tokio::spawn(async move {
-                let mut txn = store.new_txn(false).await.unwrap();
-                // Write and read contended key
-                txn.set(b"contended", format!("{}", i).as_bytes())
-                    .await
-                    .unwrap();
-                let _ = txn.get(b"contended").await.unwrap();
-                txn.commit().await.unwrap();
-                completed.fetch_add(1, Ordering::SeqCst);
-            }));
-        }
-
-        for h in handles {
-            h.await.unwrap();
-        }
-
-        assert_eq!(
-            completed.load(Ordering::SeqCst),
-            num_tasks,
-            "All 100 concurrent transactions should complete"
-        );
-
-        // Verify no transactions leaked
-        assert_eq!(
-            store.active_transaction_count(),
-            0,
-            "No active transactions should remain"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_redb_close_during_concurrent_transaction_creation() {
-        use std::sync::atomic::{AtomicUsize, Ordering};
-
-        let temp_dir = TempDir::new().unwrap();
-        let store = Arc::new(RedbStore::open(temp_dir.path().join("test.redb")).unwrap());
-
-        let completed = Arc::new(AtomicUsize::new(0));
-        let rejected = Arc::new(AtomicUsize::new(0));
-
-        // Use a barrier to synchronize all tasks to start simultaneously
-        // This ensures close() actually races with transaction creation
-        let num_txn_tasks = 50;
-        let barrier = Arc::new(tokio::sync::Barrier::new(num_txn_tasks + 1)); // +1 for close task
-
-        let mut handles = vec![];
-
-        // Spawn tasks that continuously create and complete transactions
-        for _ in 0..num_txn_tasks {
-            let store = Arc::clone(&store);
-            let completed = Arc::clone(&completed);
-            let rejected = Arc::clone(&rejected);
-            let barrier = Arc::clone(&barrier);
-
-            handles.push(tokio::spawn(async move {
-                // Wait for all tasks to be ready
-                barrier.wait().await;
-
-                for _ in 0..10 {
-                    match store.new_txn(true).await {
-                        Ok(txn) => {
-                            txn.discard();
-                            completed.fetch_add(1, Ordering::SeqCst);
-                        }
-                        Err(crate::corekv::Error::DBClosed) => {
-                            rejected.fetch_add(1, Ordering::SeqCst);
-                            return; // Stop trying after close
-                        }
-                        Err(e) => panic!("Unexpected error: {:?}", e),
-                    }
-                    // Small yield to allow close to interleave
-                    tokio::task::yield_now().await;
-                }
-            }));
-        }
-
-        // Spawn the close task that also waits at the barrier
-        let store_clone = Arc::clone(&store);
-        let barrier_clone = Arc::clone(&barrier);
-        let close_handle = tokio::spawn(async move {
-            // Wait for all tasks to be ready, then immediately close
-            barrier_clone.wait().await;
-            store_clone.close().await
-        });
-
-        // Wait for all transaction tasks
-        for handle in handles {
-            handle.await.unwrap();
-        }
-
-        // Wait for close to complete (may succeed or timeout)
-        let _close_result =
-            tokio::time::timeout(std::time::Duration::from_secs(10), close_handle).await;
-
-        // CRITICAL: Verify count is 0 regardless of close result
-        // This catches TOCTOU bugs where count goes negative or leaks
-        assert_eq!(
-            store.active_transaction_count(),
-            0,
-            "Transaction count should be 0 after all tasks complete"
-        );
-
-        // The test verifies correct behavior regardless of race outcome:
-        // - If close wins the race: many transactions will be rejected (DBClosed)
-        // - If transactions win: they complete successfully
-        // Either outcome is valid - the key invariant is that the count is 0 at the end
-        let completed_count = completed.load(Ordering::SeqCst);
-        let rejected_count = rejected.load(Ordering::SeqCst);
-        let total = completed_count + rejected_count;
-
-        // At least some activity should have happened
-        assert!(
-            total > 0,
-            "Some transactions should have been attempted (completed: {}, rejected: {})",
-            completed_count,
-            rejected_count
-        );
-    }
-
-    #[tokio::test]
-    async fn test_redb_mixed_read_write_high_contention() {
-        use std::sync::atomic::{AtomicUsize, Ordering};
-
-        let temp_dir = TempDir::new().unwrap();
-        let store = Arc::new(RedbStore::open(temp_dir.path().join("test.redb")).unwrap());
-
-        // Setup initial data
-        {
-            let mut txn = store.new_txn(false).await.unwrap();
-            for i in 0..10 {
-                txn.set(format!("key_{}", i).as_bytes(), b"initial")
-                    .await
-                    .unwrap();
-            }
-            txn.commit().await.unwrap();
-        }
-
-        let reads = Arc::new(AtomicUsize::new(0));
-        let writes = Arc::new(AtomicUsize::new(0));
-
-        let mut handles = vec![];
-
-        // 50 readers
-        for _ in 0..50 {
-            let store = Arc::clone(&store);
-            let reads = Arc::clone(&reads);
-
-            handles.push(tokio::spawn(async move {
-                for _ in 0..20 {
-                    let txn = store.new_txn(true).await.unwrap();
-                    // Read all keys
-                    for i in 0..10 {
-                        let _ = txn.get(format!("key_{}", i).as_bytes()).await.unwrap();
-                    }
-                    txn.discard();
-                    reads.fetch_add(1, Ordering::SeqCst);
-                }
-            }));
-        }
-
-        // 20 writers
-        for writer_id in 0..20 {
-            let store = Arc::clone(&store);
-            let writes = Arc::clone(&writes);
-
-            handles.push(tokio::spawn(async move {
-                for cycle in 0..10 {
-                    let mut txn = store.new_txn(false).await.unwrap();
-                    let key = format!("key_{}", cycle % 10);
-                    let value = format!("writer_{}_{}", writer_id, cycle);
-                    txn.set(key.as_bytes(), value.as_bytes()).await.unwrap();
-                    txn.commit().await.unwrap();
-                    writes.fetch_add(1, Ordering::SeqCst);
-                }
-            }));
-        }
-
-        // Wait for all to complete
-        for h in handles {
-            h.await.unwrap();
-        }
-
-        assert_eq!(
-            reads.load(Ordering::SeqCst),
-            50 * 20,
-            "All reads should complete"
-        );
-        assert_eq!(
-            writes.load(Ordering::SeqCst),
-            20 * 10,
-            "All writes should complete"
-        );
-        assert_eq!(
-            store.active_transaction_count(),
-            0,
-            "No leaked transactions"
-        );
-    }
-
-    // =========================================================================
-    // LARGE DATASET STRESS TESTS
-    // =========================================================================
-
-    #[tokio::test]
-    #[ignore] // Run with: cargo test -- --ignored (takes several seconds)
-    async fn test_redb_100k_keys_stress() {
-        let temp_dir = TempDir::new().unwrap();
-        let path = temp_dir.path().join("test.redb");
-
-        let store = RedbStore::open(&path).unwrap();
-
-        // Insert 100K keys in batches of 1000
-        for batch in 0..100 {
-            let mut txn = store.new_txn(false).await.unwrap();
-            for i in 0..1000 {
-                let key = format!("key_{:08}", batch * 1000 + i);
-                let value = vec![0xAB; 100]; // 100 bytes per value
-                txn.set(key.as_bytes(), &value).await.unwrap();
-            }
-            txn.commit().await.unwrap();
-        }
-
-        // Verify reads work with 100K keys in snapshot
-        let txn = store.new_txn(true).await.unwrap();
-        assert_eq!(
-            txn.get(b"key_00000000").await.unwrap(),
-            Some(vec![0xAB; 100]),
-            "First key should be retrievable"
-        );
-        assert_eq!(
-            txn.get(b"key_00099999").await.unwrap(),
-            Some(vec![0xAB; 100]),
-            "Last key should be retrievable"
-        );
-
-        // Test prefix iteration on large dataset
-        let opts = crate::corekv::IterOptions::new().with_prefix(b"key_00050".to_vec());
-        let mut iter = txn.iterator(opts).await.unwrap();
-        let mut count = 0;
-        while iter.next().await.unwrap().is_some() {
-            count += 1;
-        }
-        // Keys matching "key_00050*" should be key_00050000 through key_00050999
-        assert_eq!(count, 1000, "Should have 1000 keys with prefix key_00050");
-
-        txn.discard();
-        store.close().await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_redb_10k_keys_with_large_values() {
-        let temp_dir = TempDir::new().unwrap();
-        let store = RedbStore::open(temp_dir.path().join("test.redb")).unwrap();
-
-        // 10K keys with 1KB values each = ~10MB total
-        let value = vec![0xCD; 1024];
-
-        let mut txn = store.new_txn(false).await.unwrap();
-        for i in 0..10_000 {
-            let key = format!("largevalue_{:06}", i);
-            txn.set(key.as_bytes(), &value).await.unwrap();
-        }
-        txn.commit().await.unwrap();
-
-        // Verify random access
-        let txn = store.new_txn(true).await.unwrap();
-        for check in [0, 1000, 5000, 9999] {
-            let key = format!("largevalue_{:06}", check);
-            let retrieved = txn.get(key.as_bytes()).await.unwrap();
-            assert_eq!(
-                retrieved.as_ref().map(|v| v.len()),
-                Some(1024),
-                "Key {} should have 1KB value",
-                key
-            );
-        }
-        txn.discard();
-    }
-
-    // =========================================================================
-    // CALLBACK MONITORING TESTS
-    // =========================================================================
-
-    #[tokio::test]
-    async fn test_redb_callback_counts() {
-        let temp_dir = TempDir::new().unwrap();
-        let store = RedbStore::open(temp_dir.path().join("test.redb")).unwrap();
-
-        let mut txn = store.new_txn(false).await.unwrap();
-
-        // Initially all counts should be 0
-        let redb_txn = txn.as_any().downcast_ref::<RedbTxn>().unwrap();
-        let counts = redb_txn.callback_counts();
-        assert_eq!(counts.total(), 0, "Initial callback count should be 0");
-
-        // Register some callbacks
-        txn.on_success(Box::new(|| {}));
-        txn.on_success(Box::new(|| {}));
-        txn.on_error(Box::new(|| {}));
-        txn.on_discard(Box::new(|| {}));
-
-        // Check updated counts
-        let redb_txn = txn.as_any().downcast_ref::<RedbTxn>().unwrap();
-        let counts = redb_txn.callback_counts();
-        assert_eq!(counts.on_success, 2, "Should have 2 success callbacks");
-        assert_eq!(counts.on_error, 1, "Should have 1 error callback");
-        assert_eq!(counts.on_discard, 1, "Should have 1 discard callback");
-        assert_eq!(counts.total(), 4, "Total should be 4");
-
-        txn.discard();
-    }
-
-    #[tokio::test]
-    async fn test_redb_check_integrity() {
-        let temp_dir = TempDir::new().unwrap();
-        let store = RedbStore::open(temp_dir.path().join("test.redb")).unwrap();
-
-        // Empty database should pass integrity check
-        let report = store.check_integrity().unwrap();
-        assert!(
-            report.is_valid,
-            "Empty database should pass integrity check"
-        );
-        assert_eq!(report.total_keys, 0, "Empty database should have 0 keys");
-        assert_eq!(report.error_count, 0, "Empty database should have 0 errors");
-        assert!(
-            report.first_error.is_none(),
-            "Empty database should have no error message"
-        );
-
-        // Add some data
-        {
-            let mut txn = store.new_txn(false).await.unwrap();
-            for i in 0..100 {
-                let key = format!("key_{}", i);
-                let value = format!("value_{}", i);
-                txn.set(key.as_bytes(), value.as_bytes()).await.unwrap();
-            }
-            txn.commit().await.unwrap();
-        }
-
-        // Database with data should pass integrity check
-        let report = store.check_integrity().unwrap();
-        assert!(
-            report.is_valid,
-            "Database with data should pass integrity check"
-        );
-        assert_eq!(report.total_keys, 100, "Database should have 100 keys");
-        assert_eq!(report.error_count, 0, "Database should have 0 errors");
-        assert!(
-            report.first_error.is_none(),
-            "Database should have no error message"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_redb_db_path() {
-        let temp_dir = TempDir::new().unwrap();
-        let expected_path = temp_dir.path().join("mytest.redb");
-        let store = RedbStore::open(&expected_path).unwrap();
-
-        assert_eq!(
-            store.db_path(),
-            expected_path,
-            "db_path() should return the correct path"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_redb_configurable_close_timeout() {
-        use std::time::Duration;
-
-        let temp_dir = TempDir::new().unwrap();
-        let opts = RedbStoreOptions::new().with_close_timeout(Duration::from_millis(100));
-        let store = Arc::new(
-            RedbStore::open_with_options(temp_dir.path().join("test.redb"), opts).unwrap(),
-        );
-
-        // Create a transaction that we intentionally don't close
-        let _held_txn = store.new_txn(true).await.unwrap();
-
-        // Close should timeout much faster (100ms instead of default 5s)
-        let start = std::time::Instant::now();
-        let result = store.close().await;
-        let elapsed = start.elapsed();
-
-        assert!(result.is_err(), "Close should timeout");
-        assert!(
-            elapsed < Duration::from_secs(1),
-            "Close should timeout quickly with custom 100ms timeout, took {:?}",
-            elapsed
-        );
-    }
-
-    #[tokio::test]
-    async fn test_redb_callback_count() {
-        let temp_dir = TempDir::new().unwrap();
-        let store = RedbStore::open(temp_dir.path().join("test.redb")).unwrap();
-
-        let mut txn = store.new_txn(false).await.unwrap();
-
-        // Initially no callbacks
-        assert_eq!(txn.callback_count(), 0, "Should start with 0 callbacks");
-
-        // Register some callbacks
-        txn.on_success(Box::new(|| {}));
-        assert_eq!(
-            txn.callback_count(),
-            1,
-            "Should have 1 callback after on_success"
-        );
-
-        txn.on_error(Box::new(|| {}));
-        assert_eq!(
-            txn.callback_count(),
-            2,
-            "Should have 2 callbacks after on_error"
-        );
-
-        txn.on_discard(Box::new(|| {}));
-        assert_eq!(
-            txn.callback_count(),
-            3,
-            "Should have 3 callbacks after on_discard"
-        );
-
-        txn.discard();
-    }
-
-    // =========================================================================
-    // MEMORY PRESSURE STRESS TEST
-    // =========================================================================
-
-    /// Test that validates memory behavior under concurrent read transactions.
-    ///
-    /// This test creates a moderately-sized dataset and opens multiple concurrent
-    /// read transactions to verify that memory pressure is manageable.
-    ///
-    /// Memory calculation: 10K keys × 100 bytes × 20 concurrent txns = ~20MB
-    #[tokio::test]
-    async fn test_redb_memory_pressure_concurrent_snapshots() {
-        use std::sync::atomic::{AtomicUsize, Ordering};
-        use std::sync::Arc;
-
-        let temp_dir = TempDir::new().unwrap();
-
-        let store = Arc::new(RedbStore::open(temp_dir.path().join("test.redb")).unwrap());
-
-        // Setup: Create 10K keys with 100-byte values (~1MB total)
-        let value = vec![0xAB; 100];
-        {
-            let mut txn = store.new_txn(false).await.unwrap();
-            for i in 0..10_000 {
-                let key = format!("memtest_{:06}", i);
-                txn.set(key.as_bytes(), &value).await.unwrap();
-            }
-            txn.commit().await.unwrap();
-        }
-
-        // Open 20 concurrent read transactions (each snapshots the entire DB)
-        let concurrent_readers = 20;
-        let completed = Arc::new(AtomicUsize::new(0));
-        let errors = Arc::new(AtomicUsize::new(0));
-
-        let mut handles = vec![];
-        for _ in 0..concurrent_readers {
-            let store = Arc::clone(&store);
-            let completed = Arc::clone(&completed);
-            let errors = Arc::clone(&errors);
-
-            handles.push(tokio::spawn(async move {
-                match store.new_txn(true).await {
-                    Ok(txn) => {
-                        // Verify we can read data
-                        let result = txn.get(b"memtest_005000").await;
-                        if result.is_ok() && result.unwrap().is_some() {
-                            completed.fetch_add(1, Ordering::SeqCst);
-                        } else {
-                            errors.fetch_add(1, Ordering::SeqCst);
-                        }
-                        txn.discard();
-                    }
-                    Err(_) => {
-                        errors.fetch_add(1, Ordering::SeqCst);
-                    }
-                }
-            }));
-        }
-
-        for h in handles {
-            h.await.unwrap();
-        }
-
-        assert_eq!(
-            completed.load(Ordering::SeqCst),
-            concurrent_readers,
-            "All concurrent read transactions should complete successfully"
-        );
-        assert_eq!(
-            errors.load(Ordering::SeqCst),
-            0,
-            "No errors should occur during concurrent reads"
-        );
-        assert_eq!(
-            store.active_transaction_count(),
-            0,
-            "All transactions should be cleaned up"
-        );
-
-        store.close().await.unwrap();
-    }
-}
